### 2019-06-25 15f53cc
Added missing close quotation mark in paper.md
