### 2019-06-25 #1526 @kyleniemeyer
Hi @johnbachman, there was a missing close quotation mark in your paper's header that I fixed in https://github.com/indralab/adeft/pull/36, if you could merge that.

In this issue, we'll get an editor assigned to your submission, who will then assign two or more reviewers. If you have suggestions for who we might ask, either from the [JOSS reviewer pool](https://bit.ly/joss-reviewers) or from folks in your domain, that would be helpful.
