### 2019-06-26 #1526 @johnbachman
Thanks very much! We've gone through the list of reviewers and suggest the following (based on listed experience in Python, NLP, and/or computational biology):

* bagheria
* ionlights
* nhejazi
* gweissman
