### 2019-06-26 9c83f07
Add missing commas to paper.bib
