+++ paper/paper.bib
@@ -22,7 +22,7 @@
    Volume="26",
    Number="9",
    Pages="1246--1253",
-   Month="May"
+   Month="May",
    doi={10.1093/bioinformatics/btq129},
    url={https://doi.org/10.1093/bioinformatics/btq129}
    }
@@ -48,7 +48,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Volume="13",
    Number="11",
    Pages="954",
-   Month="11"
+   Month="11",
    doi={10.15252/msb.20177651},
    url={https://doi.org/10.15252/msb.20177651}
 }
