### 2019-06-26 #1526 @johnbachman
Found two missing commas in ``paper.bib``, just fixed them and pushed again.
