### 2019-06-26 #1526 @johnbachman
Fixed missing references and misc formatting.
