### 2019-06-26 #1526 @johnbachman
Proofs look good.
