### 2019-06-27 #1526 @brainstorm
@vsoch, since you have done pretty rad stuff with wordvec/NLP before, would you mind helping me out on reviewing this submission? :)
