### 2019-06-27 #1526 @vsoch
hey @brainstorm ! It's been a few years since I've looked at NLP, so I would suggest that we also find someone else that is actively practicing, and have me as a second reviewer? But this looks incredibly cool and it would be an honor! I may not have time until the weekend, but that's still pretty soon :)
