### 2019-07-09 189f3e7
Add cython optimization function for longform score
