### 2019-07-09 4f5bb16
Build cython extension in setup
