### 2019-07-16 #1526 @brainstorm
@Gweismann, would you like to assist JOSS with this review?
