### 2019-07-18 836dfdf
Convert input to C data structure
