### 2019-07-18 a990bfc
Encode input for optimization problem
