### 2019-07-18 91ff6cc
Spin conversion off into its own function
