### 2019-07-18 053e96f
Add length to candidates array
