### 2019-07-18 76491df
Write function to free memory of candidates array
