### 2019-07-19 64df87b
Free memory in 2d array correctly
