### 2019-07-19 905db98
Put relevant lengths in array at top level
