### 2019-07-19 9282c91
Begin work on stitch function to assemble input
