### 2019-07-19 e4edf05
Pass pointer to struct instead of struct
