### 2019-07-19 36da7d8
Get stitch function working correctly
