### 2019-07-19 3ce43cd
Put check function in VC temporarily
