### 2019-07-19 cfe1d01
Add more complex example to check function
