### 2019-07-19 c9a118b
Get stitch to insert wildcards
