### 2019-07-19 9594b39
Print full output with wildcards included
