+++ adeft/score/score_util.pyx
@@ -224,7 +224,8 @@ def check_convert():
     output = stitch(&candidates, perm, 3)
 
     print(output[0], output[1], output[2], output[3], output[4],
-          output[5])
+          output[5], output[6], output[7], output[8], output[9],
+          output[10], output[12])
 
     free_candidates_array(&candidates)            
 
