### 2019-07-19 1b7763a
Return pointer to struct instead of struct
