### 2019-07-19 fed2791
Use new implementation with pointer returned
