### 2019-07-19 ba24b9d
Free memory of output array in check
