### 2019-07-19 973af49
Return pointer to struct in convert_input
