### 2019-07-19 c28dc77
Update check function for new return types
