### 2019-07-22 #1526 @brainstorm
:wave: @adammichaelwood, would you like to assist JOSS with this review?
