### 2019-07-22 #1526 @adammichaelwood
Thanks for asking.
I don't have time for this at the moment.

On Sun, Jul 21, 2019, 5:13 PM Roman Valls Guimera <notifications@github.com>
wrote:

> 👋 @adammichaelwood <https://github.com/adammichaelwood>, would you like
> to assist JOSS with this review?
>
