### 2019-07-22 1b1c371
Pass C arrays to optimize instead of memoryview
