### 2019-07-22 2a4ba77
Check in failed permutation algorithm
