### 2019-07-23 3e5a129
Make reference implementation of evens speedup work
