### 2019-07-23 733057f
Put penalties into struct
