### 2019-07-23 d25cdcb
Fix a number of errors
