### 2019-07-23 52bf4a1
Update check_convert to new stitching pattern
