### 2019-07-23 74430f9
Add non-working template for perm_search
