+++ adeft/score/score_util.pyx
@@ -253,6 +253,79 @@ cdef opt_input *stitch(candidates_array *candidates,
             output.prizes.array[j+1] = 0
             j += 2
     return &output
+
+
+cdef struct perm_out:
+    double score
+
+    
+cdef perm_out *perm_search(candidates_array *candidates, int n):
+    cdef:
+        int m = n - 1
+        int *P
+        int *Pinv
+        int *D
+        int *T
+        int X, Y, Z, W
+        opt_input *current
+        results *opt_results
+
+    P = <int *> PyMem_Malloc(n * sizeof(int))
+    Pinv = <int *> PyMem_Malloc(n * sizeof(int))
+    D = <int *> PyMem_Malloc(n * sizeof(int))
+    T = <int *> PyMem_Malloc(n * sizeof(int))
+
+    cdef int i = 0
+    for i in range(n):
+        P[i] = i
+        Pinv[i] = i
+        D[i] = -1
+        T[i] = -1
+
+    current = stitch(candidates, P, n)
+    opt_results = optimize(current.x.array, current.prizes.array, penalties.array,
+                           current.x.length, penalties.length)
+                           
+    while m != 0:
+        X = Pinv[m]
+        Y = X + D[m]
+        Z = P[Y]
+        P[Y] = m
+        P[X] = Z
+        Pinv[Z] = X
+        Pinv[m] = Y
+        W = Y + D[m]
+        if W == -1 or W == n or P[W] > m:
+            D[m] = -D[m]
+            if m == n - 1:
+                if T[n-1] < 0:
+                    m = n - 2
+                    if -T[n-1] != n - 1:
+                        T[n-2] = T[n-1]
+                else:
+                    m = T[n-1] - 1
+            else:
+                T[n-1] = -(m+2)
+                if T[m] > 0:
+                    T[m+1] = T[m]
+                else:
+                    T[m+1] = m
+                    if -T[m] != m:
+                        T[m-1] = T[m]
+                m = n - 1
+        else:
+            if m != n-1:
+                T[n-1] = -m - 1
+                m = n-1
+        current = stitch(candidates, P, n)
+        opt_results = optimize(current.x.array, current.prizes.array,
+                               penalties.array, current.x.length,
+                               penalties.length)
+
+    PyMem_Free(P)
+    PyMem_Free(Pinv)
+    PyMem_Free(D)
+    PyMem_Free(T)
         
 
 def check_convert():
