### 2019-07-23 3626461
Add stub for longform_score calculator
