+++ adeft/score/score_util.pyx
@@ -5,10 +5,21 @@ from cpython.mem cimport PyMem_Malloc, PyMem_Free
 
 def check_optimize():
     cdef:
-        int x[10]
-        int y[2]
-        double prizes[10]
-        double penalties[2]
+        int x_0[10]
+        int y_0[2]
+        double prizes_0[10]
+        double penalties_0[2]
+        int_array x, y
+        double_array prizes, penalties
+
+    x.array = x_0
+    x.length = 10
+    y.array = y_0
+    y.length = 2
+    prizes.array = prizes_0
+    prizes.length = 10
+    penalties.array = penalties_0
+    penalties.length = 2
 
     a = np.array([-1, 0, -1, 1, -1, 0, -1, 1, -1, 0], dtype=np.int)
     b = np.array([0, 1], dtype=np.int)
@@ -24,7 +35,7 @@ def check_optimize():
         y[i] = b[i]
         penalties[i] = d[i]
 
-    output = optimize(x, y, prizes, penalties, 10, 2)
+    output = optimize(&x, &y, &prizes, &penalties)
     score = output.score
     indices = output.indices
     chars_matched = output.chars_matched
@@ -40,9 +51,8 @@ cdef struct results:
     int chars_matched
 
 
-cdef results *optimize(int *x, int *y,
-                       double *prizes, double *penalties,
-                       int len_x, int len_y):
+cdef results *optimize(int_array *x, int_array *y,
+                       double_array *prizes, double_array *penalties):
     """Subsequence match optimization algorithm for longform scoring
 
     Uses a dynamic programming algorithm to find optimal instance of
@@ -77,11 +87,11 @@ cdef results *optimize(int *x, int *y,
         in y that were matched in x. 
     """
     cdef:
-        unsigned int n = len_x
-        unsigned int m = len_y
+        unsigned int n = x.length
+        unsigned int m = y.length
         double possibility1, possibility
         unsigned int i, j, k
-        results output
+        results *output
 
     # Dynamic initialization of score_lookup array and traceback pointer
     # array
@@ -149,7 +159,7 @@ cdef results *optimize(int *x, int *y,
     # Trace backwards through pointer array to discover which elements of x
     # were matched and add the corresponding indices to the index array in
     # reverse order
-    i, j, k = n-1, m-1, 0
+    i, j, k = m-1, n-1, 0
     while i > 0:
         if pointers[i][j]:
             i -= 1
@@ -157,14 +167,14 @@ cdef results *optimize(int *x, int *y,
             output.indices[k] = i
             k += 1
         else:
-            i -= 1
+            j -= 1
     # Free pointer array
     for i in range(n):
         PyMem_Free(pointers[i])
     PyMem_Free(pointers)
     # Set the number of chars in y that were matched
     output.chars_matched = k
-    return &output
+    return output
 
 cdef struct int_array:
     int *array
@@ -176,24 +186,40 @@ cdef struct double_array:
 
 cdef struct candidates_array:
     int_array *array
+    int_array *y
     double_array *prizes
+    double_array *penalties
     int *cum_lengths
     int length
 
     
-cdef candidates_array *convert_input(list encoded_candidates,
-                                     list prizes):
+cdef candidates_array *convert_input(list encoded_shortform,
+                                     list encoded_candidates,
+                                     list prizes,
+                                     list penalties):
     cdef:
-        int i, j, num_candidates, m, n, cum_length
-        candidates_array candidates
[…]
