### 2019-07-24 7141123
Use native C double instead of np double
