### 2019-07-24 35fff55
Index arrays correctly now that they are in structs
