### 2019-07-24 4e5595b
Initialize y array corresonding to shortform
this was the big bug that took over an hour to fix
