### 2019-07-24 df02336
Allocate memory for output struct
