### 2019-07-24 9dd759d
Update checks that will later become tests
