### 2019-07-24 774b649
Apply inversion penalty
