### 2019-07-24 8b6ad71
Remove unneeded prototype code
