+++ adeft/score/score_util.pyx
@@ -431,71 +431,4 @@ def check_perm_search():
     return score
 
 
-@boundscheck(False)
-@wraparound(False)
-def permutations(x):
-    cdef:
-        int n = x
-        int m = n - 1
-        int *P
-        int *Pinv
-        int *D
-        int *T
-        int X, Y, Z, W
-        list out
-
-    P = <int *> PyMem_Malloc(n * sizeof(int))
-    Pinv = <int *> PyMem_Malloc(n * sizeof(int))
-    D = <int *> PyMem_Malloc(n * sizeof(int))
-    T = <int *> PyMem_Malloc(n * sizeof(int))
-
-    cdef int i = 0
-    for i in range(n):
-        P[i] = i
-        Pinv[i] = i
-        D[i] = -1
-        T[i] = -1
-
-    out = [None]*n
-    for i in range(n):
-        out[i] = P[i]
-    yield out
-    while m != 0:
-        X = Pinv[m]
-        Y = X + D[m]
-        Z = P[Y]
-        P[Y] = m
-        P[X] = Z
-        Pinv[Z] = X
-        Pinv[m] = Y
-        W = Y + D[m]
-        if W == -1 or W == n or P[W] > m:
-            D[m] = -D[m]
-            if m == n - 1:
-                if T[n-1] < 0:
-                    m = n - 2
-                    if -T[n-1] != n - 1:
-                        T[n-2] = T[n-1]
-                else:
-                    m = T[n-1] - 1
-            else:
-                T[n-1] = -(m+2)
-                if T[m] > 0:
-                    T[m+1] = T[m]
-                else:
-                    T[m+1] = m
-                    if -T[m] != m:
-                        T[m-1] = T[m]
-                m = n - 1
-        else:
-            if m != n-1:
-                T[n-1] = -m - 1
-                m = n-1
-        out = [None]*n
-        for i in range(n):
-            out[i] = P[i]
-        yield out
-
-
-
 
