### 2019-07-25 68041e2
Spin permutation generation off into its own module
