### 2019-07-25 67f3d8e
Complete major refactoring
