+++ adeft/score/score_util.pyx
@@ -15,6 +15,7 @@ def check_optimize():
         double penalties_0[2]
         int_array x, y
         double_array prizes, penalties
+        opt_results *output
 
     penalties.length = 2
 
@@ -40,31 +41,216 @@ def check_optimize():
     prizes.length = 5
     penalties.array = penalties_0
 
-    output = optimize(&x, &y, &prizes, &penalties)
+    output = make_opt_results(2)
+    optimize(&x, &y, &prizes, &penalties, output)
     score = output.score
     indices = output.indices
     chars_matched = output.chars_matched
     ind = np.empty(chars_matched, dtype=np.int)
     for i in range(chars_matched):
         ind[i] = indices[i]
+    free_opt_results(output)
     return score, ind
 
 
-cdef struct results:
+cdef struct opt_results:
     double score
     int *indices
     int chars_matched
 
 
-cdef results *optimize(int_array *x, int_array *y,
-                       double_array *prizes, double_array *penalties):
+cdef opt_results *make_opt_results(int len_y):
+    cdef opt_results *results
+    results = <opt_results *> PyMem_Malloc(sizeof(opt_results))
+    results.indices = <int *> PyMem_Malloc(len_y * sizeof(int))
+    return results
+
+
+cdef void free_opt_results(opt_results *results):
+    PyMem_Free(results.indices)
+    PyMem_Free(results)
+    return
+
+cdef struct int_array:
+    int *array
+    int length
+
+
+cdef int_array *make_int_array(int length):
+    cdef int_array *output
+    output = <int_array *> PyMem_Malloc(sizeof(int_array))
+    output.array = <int *> PyMem_Malloc(length * sizeof(int))
+    output.length = length
+    return output
+
+
+cdef void free_int_array(int_array *x):
+    PyMem_Free(x.array)
+    PyMem_Free(x)
+    return
+
+
+cdef struct double_array:
+    double *array
+    int length
+
+    
+cdef double_array *make_double_array(int length):
+    cdef double_array *output
+    output = <double_array *> PyMem_Malloc(sizeof(int_array))
+    output.array = <double *> PyMem_Malloc(length * sizeof(double))
+    output.length = length
+    return output
+
+
+cdef void free_double_array(double_array *x):
+    PyMem_Free(x.array)
+    PyMem_Free(x)
+    return
+
+
+cdef struct candidates_array:
+    int_array **array
+    int_array *y
+    double_array **prizes
+    double_array *penalties
+    int *cum_lengths
+    int length
+    double inv_penalty
+
+
+cdef candidates_array *make_candidates_array(list encoded_shortform,
+                                             list encoded_candidates,
+                                             list prizes,
+                                             list penalties,
+                                             double inv_penalty):
+    cdef:
+        int i, j, num_candidates, m, n, cum_length, k
+        candidates_array *candidates
+    n = len(encoded_candidates)
+    k = len(encoded_shortform)
+    candidates = <candidates_array *> PyMem_Malloc(sizeof(candidates_array))
+    candidates.array = <int_array **> PyMem_Malloc(n * sizeof(int_array*))
+    candidates.prizes = <double_array **> \
+        PyMem_Malloc(n * sizeof(double_array*))
+    candidates.penalties = make_double_array(k)
+    candidates.cum_lengths = <int *> PyMem_Malloc(n * sizeof(int))
+    candidates.length = n
+    candidates.inv_penalty = inv_penalty
+    candidates.y = make_int_array(k)
+    for i in range(k):
+        candidates.penalties.array[i] = penalties[i]
+        candidates.y.array[i] = encoded_shortform[i]
+    cum_length = 0
+    for i in range(n):
+        m = len(encoded_candidates[i])
+        candidates.array[i] = make_int_array(m)
[…]
