### 2019-07-25 d483ed3
Remove boundschecks and wraparound checks
