### 2019-07-25 7bb33e5
Add documentation for permutation generator
