### 2019-07-25 a1b87de
Get some basic tests working
