### 2019-07-25 b576ae4
Fix bug in dynamic programming back trace
