### 2019-07-26 fda3f9d
Put word capture prize info into candidates array
