### 2019-07-26 4e634f9
Get stitch to output word boundaries and word prizes
