### 2019-07-26 8c769f2
Update test helper in reponse to previous 2 commits
