### 2019-07-26 c89a570
Begin work on words captured prizes
