### 2019-07-29 4ee1fde
Return encoding to left to right order
