+++ adeft/score/score.py
@@ -1,7 +1,7 @@
 def encode(shortform, candidates):
     n = len(shortform)
     char_map = {}
-    encoded_shortform = [None]*n
+    encoded_shortform = []
     encoded_candidates = []
     char_indices = []
     used_tokens = []
@@ -11,17 +11,17 @@ def encode(shortform, candidates):
         if shortform[i] not in char_map:
             char_map[shortform[i]] = j
             j += 1
-        encoded_shortform[n-1-i] = char_map[shortform[i]]
-    for index, candidate in enumerate(candidates[::-1]):
+        encoded_shortform.append(char_map[shortform[i]])
+    for index, candidate in enumerate(candidates):
         if set(shortform) & set(candidate):
             m = len(candidate)
             coded = []
             indices = []
             for j in range(m):
-                c = candidate[m-j-1]
+                c = candidate[j]
                 if c in char_map:
                     coded.append(char_map[c])
-                    indices.append(m-j-1)
+                    indices.append(j)
             encoded_candidates.append([coded])
             char_indices.append(indices)
             used_tokens.append(index)
