### 2019-07-29 ee56a16
Increase prizes when multiple letters match in word
