+++ adeft/score/score_util.pyx
@@ -72,6 +72,7 @@ cdef struct candidates_array:
     int *cum_lengths
     int length
     double inv_penalty
+    double alpha
 
 
 @boundscheck(False)
@@ -81,7 +82,8 @@ cdef candidates_array *make_candidates_array(list encoded_shortform,
                                              list prizes,
                                              list penalties,
                                              list word_prizes,
-                                             double inv_penalty):
+                                             double inv_penalty,
+                                             double alpha):
     cdef:
         int i, j, num_candidates, m, n, cum_length, k
         candidates_array *candidates
@@ -96,6 +98,7 @@ cdef candidates_array *make_candidates_array(list encoded_shortform,
     candidates.cum_lengths = <int *> PyMem_Malloc(n * sizeof(int))
     candidates.length = n
     candidates.inv_penalty = inv_penalty
+    candidates.alpha = alpha
     candidates.y = make_int_array(k)
     for i in range(k):
         candidates.penalties.array[i] = penalties[i]
@@ -132,8 +135,8 @@ cdef free_candidates_array(candidates_array *candidates):
 cdef struct opt_input:
     int_array *x
     double_array *prizes
-    int_array *word_boundaries
-    double *word_prizes
+    unsigned int *word_boundaries
+    double_array *word_prizes
 
 
 cdef opt_input *make_opt_input(int n, int num_words):
@@ -141,8 +144,9 @@ cdef opt_input *make_opt_input(int n, int num_words):
     input_ = <opt_input *> PyMem_Malloc(sizeof(opt_input))
     input_.x = make_int_array(n)
     input_.prizes = make_double_array(n)
-    input_.word_boundaries = make_int_array(num_words)
-    input_.word_prizes = <double *> PyMem_Malloc(num_words * sizeof(double))
+    input_.word_boundaries = <unsigned int *> \
+        PyMem_Malloc(num_words * sizeof(unsigned int))
+    input_.word_prizes = make_double_array(num_words)
     return input_
 
 
@@ -169,13 +173,15 @@ cdef double perm_search(candidates_array *candidates, int n):
     perms = make_permuter(n)
     stitch(candidates, perms.P, n, current)
     optimize(current.x, candidates.y, current.prizes,
-             candidates.penalties, results)
+             candidates.penalties, current.word_boundaries,
+             current.word_prizes.array, candidates.alpha, results)
     best = results.score
     while perms.m != 0:
         update_permuter(perms)
         stitch(candidates, perms.P, n, current)
         optimize(current.x, candidates.y, current.prizes,
-                 candidates.penalties, results)
+                 candidates.penalties, current.word_boundaries,
+                 current.word_prizes.array, candidates.alpha, results)
         current_score = results.score * cpow(candidates.inv_penalty,
                                              perms.inversions)
         if current_score > best:
@@ -206,8 +212,8 @@ cdef void *stitch(candidates_array *candidates, int *permutation,
                 candidates.prizes[permutation[i]].array[k]
             result.prizes.array[j+1] = 0
             j += 2
-        result.word_prizes[i] = candidates.word_prizes[i]
-        result.word_boundaries.array[i] = j - 1
+        result.word_prizes.array[i] = candidates.word_prizes[i]
+        result.word_boundaries[i] = j - 1
     return result
 
 
@@ -215,7 +221,8 @@ cdef void *stitch(candidates_array *candidates, int *permutation,
 @wraparound(False)
 cdef void *optimize(int_array *x, int_array *y,
[…]
