### 2019-08-02 #1526 @gweissman
@brainstorm Thanks for thinking of me. I am unavailable until October.
