### 2019-08-02 5291945
Fix off by one error in array initialization
