### 2019-08-02 0288858
Fix off by one error when check for word boundaries
