### 2019-08-07 278bb0b
Free word_use lookup table when finished
