### 2019-08-07 3714939
Let test helper accept arbitrary input
