### 2019-08-07 0afe254
Rewrite test passing test_case to optimize
