### 2019-08-08 931631c
Refactor running of test cases
