### 2019-08-08 b059b90
Default to earlier if equal scoring match is found
