### 2019-08-08 798beac
Write more test cases for optimization algorithm
