+++ adeft/tests/test_score.py
@@ -13,6 +13,8 @@ def test_make_candidates_array():
 
 
 def test_optimize():
+    # Two words with beginning characters matching the characters in the
+    # shortform
     case1 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1],
                                  y=[0, 1],
                                  prizes=[0., 1., 0., 1., 0., 0.5, 0.],
@@ -23,7 +25,58 @@ def test_optimize():
                                  result_score=4.0,
                                  result_indices=[3, 1])
 
-    test_cases = [case1]
+    # Single word containing consecutive characters from shortform
+    case2 = OptimizationTestCase(x=[-1, 0, -1, 0, -1, 1, -1, 1, -1],
+                                 y=[0, 1],
+                                 prizes=[0., 1., 0., 0.5, 0.,
+                                         0.25, 0., 0.125, 0.],
+                                 penalties=[0.4, 0.2],
+                                 word_boundaries=[8],
+                                 word_prizes=[1],
+                                 alpha=0.5,
+                                 result_score=2.5,
+                                 result_indices=[5, 1])
+
+    # Three words, shortform matches in two places. Highest scoring match
+    # has a larger word captured prize and smaller total letter prizes
+    case3 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1, 0, -1, 1, -1],
+                                 y=[0, 1],
+                                 prizes=[0., 1., 0., 1., 0., 1., 0., 0.5,
+                                         0., 0.25, 0.],
+                                 penalties=[0.4, 0.2],
+                                 word_boundaries=[2, 4, 10],
+                                 word_prizes=[0.5, 0.5, 1.75],
+                                 alpha=0.5,
+                                 result_score=3.25,
+                                 result_indices=[9, 5])
+
+    # Three words, shortform matches in two places. Highest scoring match
+    # has larger total letter prizes and smaller word captured prizes
+    case4 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1, 0, -1, 1, -1],
+                                 y=[0, 1],
+                                 prizes=[0., 1., 0., 1., 0., 1., 0., 0.5,
+                                         0., 0.25, 0.],
+                                 penalties=[0.4, 0.2],
+                                 word_boundaries=[2, 4, 10],
+                                 word_prizes=[0.5, 0.5, 1.25],
+                                 alpha=0.5,
+                                 result_score=3.,
+                                 result_indices=[3, 1])
+
+    # Three words, shortform matches in two places with equal scores
+    case5 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1, 0, -1, 1, -1],
+                                 y=[0, 1],
+                                 prizes=[0., 1., 0., 1., 0., 1., 0., 0.5,
+                                         0., 0.25, 0.],
+                                 penalties=[0.4, 0.2],
+                                 word_boundaries=[2, 4, 10],
+                                 word_prizes=[0.5, 0.5, 1.5],
+                                 alpha=0.5,
+                                 result_score=3.,
+                                 result_indices=[3, 1])
+
+
+    test_cases = [case1, case2, case3, case4, case5]
     for case in test_cases:
         case.check_assertions()
         case.run_test()
@@ -63,5 +116,3 @@ class OptimizationTestCase(object):
         score, ind = check_optimize(self)
         assert score == self.result_score
         assert ind == self.result_indices
-                   
-        
