### 2019-08-09 2f1d152
Fix some formatting issues
