### 2019-08-09 df13961
Calculate cumulative length in reverse order
