### 2019-08-09 172a77a
Adjust stitching to reversing order of candidates
