### 2019-08-09 6021e89
Refactor candidates array test similar to opt test
