### 2019-08-09 8a86b54
Remove accidentally committed print statements
