### 2019-08-09 68e2c54
Remove defunct test for perm_search
