### 2019-08-09 5ffa96d
Make StitchTestCase a builtin C class
