### 2019-08-09 6c2c52a
Make OptimizationTestCase a C class
