### 2019-08-09 deb9008
Add second test case for stitching
