### 2019-08-16 0c1037d
Change output of optimize function
Originally returned indices in longform array matched by elements of
shortform array, but now returns scores associated with each char in
the shortform array.
