### 2019-08-16 93f44b8
Test case where a char in shortform doesn't match
