### 2019-08-16 e1c762e
Fix off by one when char in shortform doesn't match
