### 2019-08-29 #1526 @bgyori
Hi @brainstorm, is there some way we can help move the review process forward? Should we suggest outside reviewers not listed in the JOSS spreadsheet?
