### 2019-08-29 5b07b5d
Make score formula balance char and word prizes
    The optimization function has now been modified to take structs as
    input instead of what has become 11 separate parameters due to
    additional parameters added for this change
