+++ adeft/score/score_util.pyx
@@ -70,19 +70,14 @@ cdef struct candidates_array:
     double *word_prizes
     int *cum_lengths
     int length
-    double inv_penalty
-    double alpha
-
+    double inv_penalty, beta
 
 @boundscheck(False)
 @wraparound(False)
-cdef candidates_array *make_candidates_array(list encoded_shortform,
-                                             list encoded_candidates,
+cdef candidates_array *make_candidates_array(list encoded_candidates,
                                              list prizes,
                                              list penalties,
-                                             list word_prizes,
-                                             double inv_penalty,
-                                             double alpha):
+                                             list word_prizes)
     cdef:
         int i, j, num_candidates, m1,m2, n, cum_length, k
         candidates_array *candidates
@@ -92,16 +87,9 @@ cdef candidates_array *make_candidates_array(list encoded_shortform,
     candidates.array = <int_array **> PyMem_Malloc(n * sizeof(int_array*))
     candidates.prizes = <double_array **> \
         PyMem_Malloc(n * sizeof(double_array*))
-    candidates.penalties = make_double_array(k)
     candidates.word_prizes = <double *> PyMem_Malloc(n * sizeof(double))
     candidates.cum_lengths = <int *> PyMem_Malloc(n * sizeof(int))
     candidates.length = n
-    candidates.inv_penalty = inv_penalty
-    candidates.alpha = alpha
-    candidates.y = make_int_array(k)
-    for i in range(k):
-        candidates.penalties.array[i] = penalties[i]
-        candidates.y.array[i] = encoded_shortform[i]
     cum_length = 0
     for i in range(n):
         m1 = len(encoded_candidates[i])
@@ -126,6 +114,7 @@ cdef free_candidates_array(candidates_array *candidates):
     free_int_array(candidates.y)
     free_double_array(candidates.penalties)
     PyMem_Free(candidates.word_prizes)
+    PyMem_Free(candidates.W_array)
     PyMem_Free(candidates.prizes)
     PyMem_Free(candidates.array)
     PyMem_Free(candidates.cum_lengths)
@@ -156,11 +145,60 @@ cdef void free_opt_input(opt_input *input_):
     PyMem_Free(input_.word_boundaries)
     PyMem_Free(input_.word_prizes)
     PyMem_Free(input_)
-    
+
+
+cdef struct opt_params:
+    double beta, rho, W, C, inv_penalty
+
+
+cdef opt_params *make_opt_params(double beta, double rho, double C, double W,
+                                 double inv_penalty):
+    cdef opt_params *params
+    params = <opt_params *> PyMem_Malloc(sizeof(opt_params))
+    params.beta = beta
+    params.rho = rho
+    params.C = C
+    params.W = W
+    params.inv_penalty = inv_penalty
+    return params
+
+
+cdef void free_opt_params(opt_params *params):
+    PyMem_Free(params)
+
+
+cdef struct opt_shortform:
+    int_array *y
+    double_array *penalties
+
+
+cdef opt_shortform *make_opt_shortform(list encoded_shortform,
+                                       list penalties):
+    cdef:
+        opt_shortform *shortform
+        int i = 0
+        int m = len(encoded_shortform)
+    shortform = PyMem_Malloc(sizeof(opt_shortform))
+    shortform.y = make_int_array(m)
+    shortform.penalties = <double *> PyMem_Malloc(m*sizeof(double))
+    for i in range(m):
+        shortform.y[i] = encoded_shortform[i]
+        shortform.penalties[i] = penalties[i]
+    return shortform
+
+
+cdef void free_opt_shortform(opt_shortform *shortform):
+    free_int_array(shortform.y)
+    PyMem_Free(shortform.penalties)
+    PyMem_Free(shortform)
+
 
 @boundscheck(False)
[…]
