### 2019-08-30 4d19a25
Get new score weighting scheme to work
