+++ adeft/score/score_util.pyx
@@ -64,31 +64,31 @@ cdef void free_double_array(double_array *x):
 
 cdef struct candidates_array:
     int_array **array
-    int_array *y
     double_array **prizes
-    double_array *penalties
     double *word_prizes
+    double *W_array
     int *cum_lengths
     int length
-    double inv_penalty, beta
+
 
 @boundscheck(False)
 @wraparound(False)
 cdef candidates_array *make_candidates_array(list encoded_candidates,
                                              list prizes,
-                                             list penalties,
-                                             list word_prizes)
+                                             list word_prizes,
+                                             list W):
     cdef:
-        int i, j, num_candidates, m1,m2, n, cum_length, k
+        int i, j, num_candidates, m1,m2, n, cum_length
         candidates_array *candidates
+    print('***', W)
     n = len(encoded_candidates)
-    k = len(encoded_shortform)
     candidates = <candidates_array *> PyMem_Malloc(sizeof(candidates_array))
     candidates.array = <int_array **> PyMem_Malloc(n * sizeof(int_array*))
     candidates.prizes = <double_array **> \
         PyMem_Malloc(n * sizeof(double_array*))
     candidates.word_prizes = <double *> PyMem_Malloc(n * sizeof(double))
     candidates.cum_lengths = <int *> PyMem_Malloc(n * sizeof(int))
+    candidates.W_array = <double *> PyMem_Malloc(n * sizeof(double))
     candidates.length = n
     cum_length = 0
     for i in range(n):
@@ -99,6 +99,7 @@ cdef candidates_array *make_candidates_array(list encoded_candidates,
         cum_length += m2
         candidates.cum_lengths[i] = cum_length
         candidates.word_prizes[i] = word_prizes[i]
+        candidates.W_array[i] = W[i]
         for j in range(m1):
             candidates.array[i].array[j] = encoded_candidates[i][j]
             candidates.prizes[i].array[j] = prizes[i][j]
@@ -111,8 +112,6 @@ cdef free_candidates_array(candidates_array *candidates):
     for i in range(candidates.length):
          free_int_array(candidates.array[i])
          free_double_array(candidates.prizes[i])
-    free_int_array(candidates.y)
-    free_double_array(candidates.penalties)
     PyMem_Free(candidates.word_prizes)
     PyMem_Free(candidates.W_array)
     PyMem_Free(candidates.prizes)
@@ -126,6 +125,7 @@ cdef struct opt_input:
     double_array *prizes
     unsigned int *word_boundaries
     double_array *word_prizes
+    double W
 
 
 cdef opt_input *make_opt_input(int n, int num_words):
@@ -148,18 +148,14 @@ cdef void free_opt_input(opt_input *input_):
 
 
 cdef struct opt_params:
-    double beta, rho, W, C, inv_penalty
+    double beta, rho
 
 
-cdef opt_params *make_opt_params(double beta, double rho, double C, double W,
-                                 double inv_penalty):
+cdef opt_params *make_opt_params(double beta, double rho):
     cdef opt_params *params
     params = <opt_params *> PyMem_Malloc(sizeof(opt_params))
     params.beta = beta
     params.rho = rho
-    params.C = C
-    params.W = W
-    params.inv_penalty = inv_penalty
     return params
 
 
@@ -178,33 +174,39 @@ cdef opt_shortform *make_opt_shortform(list encoded_shortform,
         opt_shortform *shortform
         int i = 0
         int m = len(encoded_shortform)
-    shortform = PyMem_Malloc(sizeof(opt_shortform))
+    shortform = <opt_shortform *> PyMem_Malloc(sizeof(opt_shortform))
     shortform.y = make_int_array(m)
[…]
