### 2019-08-30 2a374a2
Update tests in response to score weighting change
