### 2019-09-03 f4ab743
Update perm_search for new input type
