### 2019-09-03 64553a6
Add temporary test for perm_search back
