### 2019-09-03 48fa7a3
Update optimize tests for new scoring scheme
