### 2019-09-03 c11f8cb
Remove print statements
