### 2019-09-03 6ed304b
Ensure number of inversions is initialized
