### 2019-09-03 d76c6df
Create C class for perm_search tests
