### 2019-09-03 b7b15a0
Spin out score test cases into separate module
