+++ adeft/score/score.pyx
@@ -7,11 +7,6 @@ from adeft.score.permutations cimport permuter, make_permuter, \
     free_permuter, update_permuter
 
 
-cdef struct opt_results:
-    double score
-    double *char_scores
-
-
 cdef opt_results *make_opt_results(int len_y):
     cdef opt_results *results
     results = <opt_results *> PyMem_Malloc(sizeof(opt_results))
@@ -23,10 +18,6 @@ cdef void free_opt_results(opt_results *results):
     PyMem_Free(results.char_scores)
     PyMem_Free(results)
     return
-
-cdef struct int_array:
-    int *array
-    int length
 
 
 cdef int_array *make_int_array(int length):
@@ -43,11 +34,6 @@ cdef void free_int_array(int_array *x):
     return
 
 
-cdef struct double_array:
-    double *array
-    int length
-
-    
 cdef double_array *make_double_array(int length):
     cdef double_array *output
     output = <double_array *> PyMem_Malloc(sizeof(int_array))
@@ -60,15 +46,6 @@ cdef void free_double_array(double_array *x):
     PyMem_Free(x.array)
     PyMem_Free(x)
     return
-
-
-cdef struct candidates_array:
-    int_array **array
-    double_array **prizes
-    double *word_prizes
-    double *W_array
-    int *cum_lengths
-    int length
 
 
 @boundscheck(False)
@@ -105,7 +82,7 @@ cdef candidates_array *make_candidates_array(list encoded_candidates,
     return candidates
 
 
-cdef free_candidates_array(candidates_array *candidates):
+cdef void free_candidates_array(candidates_array *candidates):
     cdef:
         int i, j
     for i in range(candidates.length):
@@ -117,14 +94,6 @@ cdef free_candidates_array(candidates_array *candidates):
     PyMem_Free(candidates.array)
     PyMem_Free(candidates.cum_lengths)
     PyMem_Free(candidates)
-
-
-cdef struct opt_input:
-    int_array *x
-    double_array *prizes
-    unsigned int *word_boundaries
-    double_array *word_prizes
-    double W
 
 
 cdef opt_input *make_opt_input(int n, int num_words):
@@ -146,10 +115,6 @@ cdef void free_opt_input(opt_input *input_):
     PyMem_Free(input_)
 
 
-cdef struct opt_params:
-    double beta, rho
-
-
 cdef opt_params *make_opt_params(double beta, double rho):
     cdef opt_params *params
     params = <opt_params *> PyMem_Malloc(sizeof(opt_params))
@@ -160,11 +125,6 @@ cdef opt_params *make_opt_params(double beta, double rho):
 
 cdef void free_opt_params(opt_params *params):
     PyMem_Free(params)
-
-
-cdef struct opt_shortform:
-    int_array *y
-    double_array *penalties
 
 
 cdef opt_shortform *make_opt_shortform(list encoded_shortform,
@@ -256,10 +216,8 @@ cdef void *stitch(candidates_array *candidates, int *permutation,
 
 @boundscheck(False)
 @wraparound(False)
-cdef void *optimize(opt_input *input_,
-                    opt_shortform *shortform,
-                    opt_params *params,
-                    opt_results *output):
+cdef void *optimize(opt_input *input_, opt_shortform *shortform,
+                    opt_params *params, opt_results *output):
     """Subsequence match optimization algorithm for longform scoring
 
     Uses a dynamic programming algorithm to find optimal instance of
@@ -402,163 +360,3 @@ cdef void *optimize(opt_input *input_,
     PyMem_Free(pointers)
     return output
 
-
-cdef struct perm_out:
-    double score
-
-
-cdef class StitchTestCase:
-    """Test construction of candidates array and stitching"""
-    cdef:
[…]
