### 2019-09-03 d76f758
Add init for LongformScore
