### 2019-09-03 b459bfd
Use C types as attributes in LongformScorer
