### 2019-09-03 b8db2c7
Add C method to get word score
