### 2019-09-04 f389467
Use weight geomean to merge char and word penalties
