### 2019-09-04 6f2a467
Take absolute value when comparing scores
