+++ adeft/tests/util.pyx
@@ -90,7 +90,6 @@ cdef class PermSearchTestCase:
         params = make_opt_params(self.beta, self.rho)
         score = perm_search(candidates, shortform, params, self.inv_penalty,
                             self.len_perm)
-        assert (score - self.result_score) < 1e-16
         assert abs(score - self.result_score) < 1e-16
 
 
