### 2019-09-04 39c8fd8
Get perm_search test working
