### 2019-09-04 148cb15
Encode candidate longforms in LongformScorer
