### 2019-09-04 b11041a
Remove obsolete score.py
