### 2019-09-04 4c01810
Check in broken score implementation
    The score method of LongformScorer is currently broken. It
    Segfaults when called. There are more important things for me to
    work on now, but I will come back to this and fix it when time permits.
