### 2019-09-04 4f96e94
Attempt to debug Segfault
    This is nonworking code. There are print statements littered about
