### 2019-09-05 #1526 @brainstorm
This is great, thanks @GullyAPCBurns and @bgyori... if @gbader is up for the task, we can get this started (respecting @vsoch preference to defer this review to more NLP-centric pros).
