### 2019-09-05 #1708 @whedon
[REVIEW]: Adeft: Acromine-based Disambiguation of Entities from Text with applications to the biomedical literature
**Submitting author:** @johnbachman (<a href="http://orcid.org/0000-0001-6095-2466">John Bachman</a>)
**Repository:** <a href="https://github.com/indralab/adeft" target ="_blank">https://github.com/indralab/adeft</a>
**Version:** 0.5.3
