### 2019-09-05 #1526 @brainstorm
Alright, off we go!: https://github.com/openjournals/joss-reviews/issues/1708
