### 2019-09-05 4b87812
Add test case for mysteriously broken example
