+++ adeft/tests/test_score.py
@@ -139,7 +139,30 @@ def test_optimize():
                                  result_score=(2/5)**(3/4)*(2/3)**(1/4),
                                  result_char_scores=[1., -0.2])
 
-    test_cases = [case1, case2, case3, case4, case5, case6]
+    # This case fails for mysterious reasons
+    case7 = OptimizationTestCase(x=[-1, 4, -1, 3, -1, 3, -1, 4, -1, 1, -1,
+                                     0, -1, 1, -1, 2, -1, 1, -1, 4, -1, 0, -1,
+                                     4, -1, 0, -1, 1, -1, 3, -1, 4, -1, 2, -1,
+                                     1, -1, 3, -1, 4, -1, 1, -1, 2, -1],
+                                 y=[0, 1, 2, 3, 4],
+                                 prizes=[0.0, 1.0, 0.0, 0.00390625, 0.0, 1.0,
+                                         0.0, 0.25, 0.0, 0.03125, 0.0,
+                                         0.015625, 0.0, 0.0078125, 0.0, 1.0,
+                                         0.0, 0.25, 0.0, 0.125, 0.0, 0.03125,
+                                         0.0, 0.0078125, 0.0, 1.0, 0.0, 0.5,
+                                         0.0, 0.03125, 0.0, 0.015625, 0.0,
+                                         0.001953125, 0.0, 1.0, 0.0, 0.03125,
+                                         0.0, 1.0, 0.0, 0.5, 0.0, 0.25, 0.0],
+                                 penalties=[0.4**i for i in range(5)],
+                                 word_boundaries=[4, 14, 24, 34, 38, 44],
+                                 word_prizes=[1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+                                 beta=0.45,
+                                 rho=0.6,
+                                 W=6.0,
+                                 result_score=0.5,
+                                 result_char_scores=[0.2, 0.2, 0.2, 0.2, 0.2])
+
+    test_cases = [case1, case2, case3, case4, case5, case6, case7]
     for case in test_cases:
         case.check_assertions()
         case.run_test()
