### 2019-09-05 bb393b1
Change default missed char penalty algorithm
