### 2019-09-05 81d44c9
Fix bug for cases where previous best score is -inf
