+++ adeft/score/score.pyx
@@ -260,29 +260,17 @@ cdef double perm_search(candidates_array *candidates,
     optimize(current, shortform, params, results)
     best = results.score
     while perms.m != 0:
-        print(perms.m)
-        print([perms.P[i] for i in range(n)])
         update_permuter(perms)
         stitch(candidates, perms.P, n, current)
         perm = [perms.P[i] for i in range(n)]
-        broken = False
-        if perm == [5, 4, 0, 3, 1, 2]:
-            broken = True
         optimize(current, shortform, params, results)
-        if broken:
-            print('aa')
         current_score = results.score * cpow(inv_penalty,
                                              perms.inversions)
-        if broken:
-            print('bb')
         if current_score > best:
             best = current_score
-        if broken:
-            print('cc')
     free_permuter(perms)
     free_opt_results(results)
     free_opt_input(current)
-    print('dd')
     return best
 
 
@@ -386,22 +374,11 @@ cdef void *optimize(opt_input *input_, opt_shortform *shortform,
             char_scores[i][j] = 0
             word_scores[i][j] = 0
             word_use[i][j] = 0
-    print('&&&')
-    x_list = [input_.x.array[i] for i in range(n)]
-    prizes = [input_.prizes.array[i] for i in range(n)]
+    x_list = [input_.x.array[s] for s in range(n)]
+    prizes = [input_.prizes.array[s] for s in range(n)]
     num_words = input_.word_prizes.length
-    word_boundaries = [input_.word_boundaries[i] for i in range(num_words)]
-    word_prizes = [input_.word_prizes.array[i] for i in range(num_words)]
-    print('&&&')
-    broken = False
-    if x_list == [-1, 4, -1, 3, -1, 3, -1, 4, -1, 1, -1, 0, -1, 1,
-                  -1, 2, -1, 1, -1, 4, -1, 0, -1, 4, -1, 0, -1, 1,
-                  -1, 3, -1, 4, -1, 2, -1, 1, -1, 3, -1, 4, -1, 1, -1, 2, -1]:
-        broken = True
-        print('prizes', prizes)
-        print('word_boundaries', word_boundaries)
-        print('word_prizes', word_prizes)
-        print('W', input_.W)
+    word_boundaries = [input_.word_boundaries[s] for s in range(num_words)]
+    word_prizes = [input_.word_prizes.array[s] for s in range(num_words)]
     # Main loop
     k = 0
     for i in range(1, n+1):
@@ -419,10 +396,13 @@ cdef void *optimize(opt_input *input_, opt_shortform *shortform,
                 char_score = char_scores[i-1][j-1]
                 char_score += input_.prizes.array[i-1]/cpow(params.beta,
                                                             word_use[i-1][j-1])
+                if char_score < 0.0:
+                    char_score = 0.0
                 word_score = word_scores[i-1][j-1] + w
                 possibility2 = (cpow(char_score/m, params.rho) *
                                 cpow(word_score/input_.W, (1-params.rho)))
-                if possibility2 > possibility1:
+                if (score_lookup[i-1][j-1] != -1e20 and 
+                    possibility2 > possibility1):
                     score_lookup[i][j] = possibility2
                     char_scores[i][j] = char_score
                     word_scores[i][j] = word_score
@@ -449,7 +429,8 @@ cdef void *optimize(opt_input *input_, opt_shortform *shortform,
                 word_score = word_scores[i-1][j-1]
                 possibility2 = (cpow(char_score/m, params.rho) *
                                 cpow(word_score/input_.W, 1-params.rho))
-                if possibility2 > possibility1:
+                if (score_lookup[i-1][j-1] != -1e20 and
+                    possibility2 > possibility1):
[…]
