### 2019-09-05 8609a98
Add tests for currently failing case
