### 2019-09-05 fc590d7
Fix word_use not updating when no possible match
