### 2019-09-06 84feda6
Change to more sensible default parameter values
