### 2019-09-06 f626a86
Change test for negative infinity out of paranoia
