### 2019-09-06 c1a8141
Move count to be implemented outside inner loop
    We must track which word we are in. Everytime we reach a word
    boundary, the count should increase by one. Since the increment
    was accidentally put inside the inner loop, the count grew too
    large. Since C arrays do not check for buffer overflows, the
    program ran even though k exceeded the size of the word_boundaries array.
