### 2019-09-07 #1526 @GullyBurns
Question about conflict of interest: I worked in the same DARPA program as
the authors and entered into collaborative discussions although we never
published or actively worked together. Just wanted to check that this never
rose to the level of a COI?

On Wed, Sep 4, 2019 at 7:36 PM Roman Valls Guimera <notifications@github.com>
wrote:

> Alright, off we go!: #1708
> <https://github.com/openjournals/joss-reviews/issues/1708>
>
