### 2019-09-08 #1526 @brainstorm
@GullyBurns The JOSS editorial team reckons this does not qualify as a COI, so thanks for raising it up but you are free to proceed, no worries.
