### 2019-09-09 dbb19fc
Fix encoding of longforms
    The case where some tokens do not contain characters from the
    shortform was not correctly handled due to misindexing
