### 2019-09-09 #1526 @GullyBurns
Great!

Could we switch this process to @GullyAPCBurns (my unaffiliated github
account?)

On Sat, Sep 7, 2019 at 6:58 PM Roman Valls Guimera <notifications@github.com>
wrote:

> @GullyBurns <https://github.com/GullyBurns> The JOSS editorial team
> reckons this does not qualify as a COI, so thanks for raising it up but you
> are free to proceed, no worries.
>
