### 2019-09-10 50c7835
Remove lists generated for print statements
