### 2019-09-10 e61ac00
Pad out array with wildcards when too short
