### 2019-09-10 782f51f
Add comments explaining lines at word boundaries
