### 2019-09-13 63ce175
Begin work on returning opt_results from perm_search
