### 2019-09-13 e362e60
Fix bug by setting output after first opt problem
