### 2019-09-13 7afa47c
Fix bug due to incorrect variable being compared
