### 2019-09-16 #1708 @brainstorm
@GullyAPCburns @gbader, please go ahead and review away ;)

https://joss.readthedocs.io/en/latest/review_checklist.html
