### 2019-09-17 f666dde
Add function to probe if new token is worth adding
