### 2019-09-17 9b08d05
Begin work on early stopping criterion
