### 2019-09-19 fb32b5d
Check in unfinished attempts to add early stopping
