+++ adeft/score/score.pyx
@@ -44,8 +44,8 @@ cdef class LongformScorer:
         else:
             self.penalties = [gamma*delta**i
                               for i in range(self.len_shortform)]
-        self.shortform_c = make_opt_shortform(encoded_shortform,
-                                              self.penalties)
+        self.shortform_c = create_shortform(encoded_shortform,
+                                            self.penalties)
         self.params_c = make_opt_params(beta, rho)
         if word_scores is None:
             self.word_scores = {}
@@ -90,44 +90,100 @@ cdef class LongformScorer:
                         
     def score(self, candidates):
         cdef:
-            int i, j, n
-            double current_score, best_score, w, W
-            double_array best_char_scores, upper_bound
+            int i, j, k, n
+            double current_score, best_score, w, W, ub_char_scores
+            double ub_word_scores
+            double_array *best_char_scores
+            double_array *previous_word_scores
             list scores
             candidates_array *candidates_c
             opt_results *results
             opt_results *probe_results
-        scores = []
         results = make_opt_results(self.len_shortform)
         probe_results = make_opt_results(self.len_shortform)
         candidates_c = self.process_candidates(candidates)
         n = candidates_c.length
+        scores = [None]*n
         best_score = -1.0
         best_char_scores = make_double_array(self.len_shortform)
-        for j in range(self.len_shortform):
-            best_char_scores[i] = -1e20
+        for i in range(self.len_shortform):
+            best_char_scores.array[i] = -1e20
         for i in range(1, n + 1):
-            upper_bound = 0.0
+            ub_char_scores = 0.0
             probe(candidates_c.array[n - i], candidates_c.prizes[n - i],
-                  self.shortform_c, self.params_c.beta, probe_results)
+                  self.shortform_c.y, self.params_c.beta,
+                  probe_results)
+            print('***')
+            return
             for j in range(self.len_shortform):
-                if probe_results.char_score[j] > best_char_scores[j]:
-                    upper_bound += probe_results.char_score[j]
+                print('probe', probe_results.char_scores[j])
+                print('e', ub_char_scores)
+                if probe_results.char_scores[j] > best_char_scores.array[j]:
+                    print('f', ub_char_scores)
+                    ub_char_scores += probe_results.char_scores[j]
                 else:
-                    upper_bound += best_char_scores[j]
+                    print('g', ub_char_scores)
+                    ub_char_scores += best_char_scores.array[j]
+            print('****')
+            if i > 1:
+                previous_word_scores = make_double_array(i)
+                for k in range(i-1):
+                    previous_word_scores.array[k] = \
+                        candidates_c.word_prizes[n-k-1]
+                ub_word_scores = opt_selection(previous_word_scores,
+                                               self.len_shortform-1)
+                free_double_array(previous_word_scores)
+            else:
+                ub_word_scores = 0.0
+            ub_char_scores = ub_char_scores/self.len_shortform
+            W = candidates_c.W_array[i-1]
+            w = candidates_c.word_prizes[n-i]
+            ub_word_scores = (w + ub_word_scores)/W
+            upper_bound = (cpow(ub_char_scores, self.params_c.rho) *
+                           cpow(ub_word_scores, 1-self.params_c.rho))
+            if upper_bound < best_score:
+                scores[i-1] = scores[i-2]*(W - w)/W
+                continue
             perm_search(candidates_c, self.shortform_c,
[…]
