### 2019-09-24 f7ac104
Refactor model dumping/loading in classifier
