### 2019-09-24 d2f8710
Fix error in probe when shortform longer than token
