+++ doc/conf.py
@@ -24,9 +24,9 @@ copyright = '2019, Albert Steppi'
 author = 'Albert Steppi'
 
 # The short X.Y version
-version = '0.3'
+version = '0.4'
 # The full version, including alpha/beta/rc tags
-release = '0.3.0'
+release = '0.4.0'
 
 
 # -- General configuration ---------------------------------------------------
