### 2019-09-25 246b1be
Fix array index bugs that lead to segfaults
    Score function does not produce correct results yet
