### 2019-09-25 e86404b
Fix off by one error that lead to incorrect score
