### 2019-09-25 6678d64
Remove bogus line that was part of print statement
