### 2019-09-25 56f3b1d
Give proper return value for longform score
    Returns tuple of (best longform, score for best longform, and list of
    all candidate longforms with their scores)
