### 2019-09-25 5fc50b3
Add set of stopwords to nlp submodule
