### 2019-09-25 45e818a
Add smaller word prize for stopwords by default
