### 2019-09-26 #1708 @gbader
Hello. I completed the review. The paper and software look good.  My graduate student @JohnGiorgi helped review the software. We were able to get it to work and it successfully disambiguated “ER” (endoplasmic reticulum) in some test text.
