### 2019-09-26 967b9ee
Handle case where no tokens overlap with shortform
