### 2019-09-26 a8fb5d7
Fix indexing bug and make input type of W list
