### 2019-09-27 1d55ca1
Use word_prize even for unused leading words
