### 2019-09-27 fd8e83b
Ensure out generated before memory freed
