### 2019-09-27 bcf0e67
Remove senseless numpy dependency
