### 2019-09-27 e2edbb1
Fix bug due to not modifying input inplace as expect
