### 2019-09-27 4eb5f60
Get rid of unnecessary white space
