### 2019-09-27 d48e5cf
Remove extra newline
