### 2019-09-27 34c50a8
Change a few formatting issues
