### 2019-09-27 29d0f52
Create base class for recognizer
