### 2019-09-27 9a91bab
Add missing return statement
