### 2019-09-27 b52e919
Change conditional order to put return at end
