### 2019-09-27 6bfd4b4
Build working recognizer based on string matching
