### 2019-09-27 f017e78
Remove unnecessary list creation
