### 2019-09-27 5d07b33
Reinstate bounds checks where speed is less critical
