### 2019-09-27 5f16c1f
Rewrite test in response to change in input type
