### 2019-09-27 6726777
Add new Docstrings to adeft.recognize
