### 2019-09-27 693bbce
Begin Docstring for AdeftLongformScorer
