### 2019-09-27 0274ff9
Change name of class init __init__
