### 2019-09-27 251c206
Flesh out Docstring for AdeftLongformScorer
