### 2019-09-27 8356aa3
Remove outdated info from Docstring
