### 2019-09-29 7ea0017
Change class name and add to Docstring
