+++ adeft/recognize.py
@@ -248,8 +248,11 @@ class AdeftRecognizer(BaseRecognizer):
         return self.grounding_map[longform]
 
 
-class StringRecognizer(BaseRecognizer):
+class OneShotRecognizer(BaseRecognizer):
     """Identify longform expansions using subsequence matching
+
+    Uses a string matching algorithm to determine longform boundaries
+    for a defining pattern for only a single text.
 
     Parameters
     ----------
