### 2019-09-29 4c9d44b
Update DocString
