### 2019-09-29 da41e44
Adjust default parameters
