### 2019-09-29 5492ce7
Adjust how inversions are penalized
