+++ adeft/score/score.pyx
@@ -21,7 +21,7 @@ cdef class AdeftLongformScorer:
 
     A character based score and a token based score are each normalized to
     range within [0, 1] and combined using a weighted geometric mean with
-    weight rho attached to the character based score and weight 1 - rho
+    weight lambda_ attached to the character based score and weight 1 - lambda_
     attached to the token based score. A dynamic programming algorithm is
     used to solve the core optimization problem.
 
@@ -32,7 +32,7 @@ cdef class AdeftLongformScorer:
     a candidate, all permutations of its tokens are considered as potential
     matches to the shortform. This allows the algorithm to identify cases
     such as beta-2 adrenergic receptor (ADRB2). A multiplicative penalty,
-    inv_penalty, is applied for each inversion of the permutation. To avoid
+    rho, is applied for each inversion of the permutation. To avoid
     factorial complexity, a subproblem is solved at each step to determine
     if the best score seen so far could be improved by considering the
     next candidate.
@@ -67,16 +67,16 @@ cdef class AdeftLongformScorer:
         Penalties for additional characters in shortform decay exponentially
         at rate delta
         Default: 0.9
-    rho : double
+    lambda_ : double
         Real value in [0, 1]
         Weighting for character based scoring vs token based scoring. Larger
-        values of rho correspond to more importance being given to character
-        matching
-        Default: 0.6
-    inv_penalty : double
+        values of lambda_ correspond to more importance being given to
+        character matching
+        Default: 0.7
+    rho : double
         Multiplicative penalty for number of inversions in permutation of
         tokens If trying a match with a permution P of the tokens in a
-        candidate, multiply score by inv_penalty**inv where inv is the number
+        candidate, multiply score by rho**inv where inv is the number
         of inversions in P
         Default: 0.9
     word_scores : dict
@@ -87,15 +87,15 @@ cdef class AdeftLongformScorer:
     cdef:
         public str shortform
         public list penalties
-        public double alpha, beta, gamma, delta, rho, inv_penalty
+        public double alpha, beta, gamma, delta, lambda_, rho
         public dict word_scores
         int len_shortform
         dict char_map
         opt_shortform *shortform_c
         opt_params *params_c
 
-    def __init__(self, shortform, penalties=None, alpha=0.2, beta=0.25,
-                 gamma=1.0, delta=0.8, rho=0.7, inv_penalty=0.95,
+    def __init__(self, shortform, penalties=None, alpha=0.4, beta=0.4,
+                 gamma=1.0, delta=0.8, lambda_=0.65, rho=0.8,
                  word_scores=None):
         self.shortform = shortform.lower()
         self.len_shortform = len(shortform)
@@ -103,9 +103,9 @@ cdef class AdeftLongformScorer:
         self.beta = beta
         self.gamma = gamma
         self.delta = delta
-        self.inv_penalty = inv_penalty
         self.rho = rho
-        self.params_c = make_opt_params(beta, rho)
+        self.lambda_ = lambda_
+        self.params_c = make_opt_params(beta, lambda_)
         # Encode shortform chars as integers and build map for encoding
         # longform candidates
         self.char_map = {}
@@ -123,7 +123,7 @@ cdef class AdeftLongformScorer:
                               for i in range(self.len_shortform)]
         self.shortform_c = create_shortform(encoded_shortform,
                                             self.penalties)
-        self.params_c = make_opt_params(beta, rho)
+        self.params_c = make_opt_params(beta, lambda_)
         if word_scores is None:
             self.word_scores = {word: 0.2 for word in stopwords}
         else:
@@ -190,7 +190,7 @@ cdef class AdeftLongformScorer:
             else:
[…]
