### 2019-09-29 c2842e2
Write tests for one shot recognizer
