### 2019-09-30 89ac05a
Commit broken attempt to modify character scoring
