+++ adeft/score/score.pyx
@@ -87,15 +87,15 @@ cdef class AdeftLongformScorer:
     cdef:
         public str shortform
         public list penalties
-        public double alpha, beta, gamma, delta, lambda_, rho
+        public double alpha, beta, gamma, delta, epsilon, lambda_, rho
         public dict word_scores
         int len_shortform
         dict char_map
         opt_shortform *shortform_c
         opt_params *params_c
 
-    def __init__(self, shortform, penalties=None, alpha=0.4, beta=0.4,
-                 gamma=1.0, delta=0.8, lambda_=0.65, rho=0.8,
+    def __init__(self, shortform, penalties=None, alpha=0.5, beta=0.9,
+                 gamma=0.95, delta=0.4, epsilon=0.8, lambda_=0.65, rho=0.8,
                  word_scores=None):
         self.shortform = shortform.lower()
         self.len_shortform = len(shortform)
@@ -103,9 +103,10 @@ cdef class AdeftLongformScorer:
         self.beta = beta
         self.gamma = gamma
         self.delta = delta
+        self.epsilon = epsilon
         self.rho = rho
         self.lambda_ = lambda_
-        self.params_c = make_opt_params(beta, lambda_)
+        self.params_c = make_opt_params(alpha, beta, gamma, lambda_)
         # Encode shortform chars as integers and build map for encoding
         # longform candidates
         self.char_map = {}
@@ -119,11 +120,11 @@ cdef class AdeftLongformScorer:
         if penalties is not None:
             self.penalties = penalties
         else:
-            self.penalties = [gamma*delta**i
+            self.penalties = [delta*epsilon**i
                               for i in range(self.len_shortform)]
         self.shortform_c = create_shortform(encoded_shortform,
                                             self.penalties)
-        self.params_c = make_opt_params(beta, lambda_)
+        self.params_c = make_opt_params(alpha, beta, gamma, lambda_)
         if word_scores is None:
             self.word_scores = {word: 0.2 for word in stopwords}
         else:
@@ -139,25 +140,20 @@ cdef class AdeftLongformScorer:
         cdef:
             double word_score
             str token
-            list encoded_candidates, coded, prizes,
-            list token_prizes, word_prizes, W
+            list encoded_candidates, coded, word_prizes, W
             int n, i, j
         encoded_candidates = []
-        prizes = []
         word_prizes = []
         n = len(candidates)
         for i in range(n):
             coded = []
-            token_prizes = []
             token = candidates[i].lower()
             m = len(token)
             for j in range(m):
                 if token[j] in self.char_map:
                     coded.append(self.char_map[token[j]])
-                    token_prizes.append(self.alpha**j)
             if coded:
                 encoded_candidates.append(coded)
-                prizes.append(token_prizes)
                 word_score = self.get_word_score(token)
                 word_prizes.append(word_score)
         if not encoded_candidates:
@@ -165,7 +161,7 @@ cdef class AdeftLongformScorer:
         W = [word_prizes[-1]]
         for i in range(1, len(word_prizes)):
             W.append(W[i-1] + word_prizes[-i])
-        return (encoded_candidates, prizes, word_prizes, W)
+        return (encoded_candidates, word_prizes, W)
 
     cdef tuple get_score_results(self,
                                  list candidates,
@@ -205,20 +201,20 @@ cdef class AdeftLongformScorer:
             double ub_word_scores
             double_array *best_char_scores
             double_array *previous_word_scores
-            list scores, encoded_candidates, prizes, word_prizes, W_array
+            list scores, encoded_candidates, word_prizes, W_array
[…]
