### 2019-09-30 e67e4d4
Fix one bug that had been introduced
