### 2019-10-01 6202cb8
Sort out incorrect char_prizes
