### 2019-10-02 9e6ffd9
Pass indices to optimizer for calculating prizes
