+++ adeft/score/score.pyx
@@ -94,8 +94,8 @@ cdef class AdeftLongformScorer:
         opt_shortform *shortform_c
         opt_params *params_c
 
-    def __init__(self, shortform, penalties=None, alpha=0.5, beta=0.9,
-                 gamma=0.95, delta=0.4, epsilon=0.8, lambda_=0.65, rho=0.8,
+    def __init__(self, shortform, penalties=None, alpha=0.2, beta=0.9,
+                 gamma=0.95, delta=1.0, epsilon=0.4, lambda_=0.72, rho=0.95,
                  word_scores=None):
         self.shortform = shortform.lower()
         self.len_shortform = len(shortform)
@@ -140,20 +140,25 @@ cdef class AdeftLongformScorer:
         cdef:
             double word_score
             str token
-            list encoded_candidates, coded, word_prizes, W
+            list encoded_candidates, indices, encoded_token
+            list token_indices, word_prizes, W
             int n, i, j
         encoded_candidates = []
+        indices = []
         word_prizes = []
         n = len(candidates)
         for i in range(n):
-            coded = []
+            encoded_token = []
+            token_indices = []
             token = candidates[i].lower()
             m = len(token)
             for j in range(m):
                 if token[j] in self.char_map:
-                    coded.append(self.char_map[token[j]])
-            if coded:
-                encoded_candidates.append(coded)
+                    encoded_token.append(self.char_map[token[j]])
+                    token_indices.append(j)
+            if encoded_token:
+                encoded_candidates.append(encoded_token)
+                indices.append(token_indices)
                 word_score = self.get_word_score(token)
                 word_prizes.append(word_score)
         if not encoded_candidates:
@@ -161,7 +166,7 @@ cdef class AdeftLongformScorer:
         W = [word_prizes[-1]]
         for i in range(1, len(word_prizes)):
             W.append(W[i-1] + word_prizes[-i])
-        return (encoded_candidates, word_prizes, W)
+        return (encoded_candidates, indices, word_prizes, W)
 
     cdef tuple get_score_results(self,
                                  list candidates,
@@ -201,19 +206,20 @@ cdef class AdeftLongformScorer:
             double ub_word_scores
             double_array *best_char_scores
             double_array *previous_word_scores
-            list scores, encoded_candidates, word_prizes, W_array
+            list scores, encoded_candidates, indices, word_prizes, W_array
             tuple out
             candidates_array *candidates_c
             opt_results *results
             opt_results *probe_results
         results = make_opt_results(self.len_shortform)
         probe_results = make_opt_results(self.len_shortform)
-        encoded_candidates, word_prizes, W_array = \
+        encoded_candidates, indices, word_prizes, W_array = \
             self.process_candidates(candidates)
         if not encoded_candidates:
             return self.get_score_results(candidates, [0.0]*len(candidates),
                                           [0.0]*len(candidates))
         candidates_c = make_candidates_array(encoded_candidates,
+                                             indices,
                                              word_prizes,
                                              W_array)
         n = candidates_c.length
@@ -224,7 +230,8 @@ cdef class AdeftLongformScorer:
             best_char_scores.array[i] = -1e20
         for i in range(1, n + 1):
             ub_char_scores = 0.0
-            probe(candidates_c.array[n - i], self.shortform_c.y,
+            probe(candidates_c.array[n - i], candidates_c.indices[n - i],
+                  self.shortform_c.y,
[…]
