### 2019-10-03 1de737d
Add optimizations
Use upper bound to check if it is even worthwhile to try permutations
of tokens. If permutation search is found to be necessary, do not run
optimization algorithm if number of permutations is too large to
improve upon the current best score
