+++ adeft/score/score.pyx
@@ -1,4 +1,6 @@
 from libc.math cimport pow as cpow
+cdef extern from 'limits.h':
+    int INT_MAX
 from cython cimport boundscheck, wraparound
 from cpython.mem cimport PyMem_Malloc, PyMem_Free
 
@@ -94,8 +96,8 @@ cdef class AdeftLongformScorer:
         opt_shortform *shortform_c
         opt_params *params_c
 
-    def __init__(self, shortform, penalties=None, alpha=0.2, beta=0.9,
-                 gamma=0.95, delta=1.0, epsilon=0.4, lambda_=0.72, rho=0.95,
+    def __init__(self, shortform, penalties=None, alpha=0.2, beta=0.85,
+                 gamma=0.9, delta=1.0, epsilon=0.4, lambda_=0.6, rho=0.95,
                  word_scores=None):
         self.shortform = shortform.lower()
         self.len_shortform = len(shortform)
@@ -201,7 +203,7 @@ cdef class AdeftLongformScorer:
 
     def score(self, candidates):
         cdef:
-            int i, j, k, n
+            int i, j, k, n, max_inversions
             double current_score, best_score, w, W, ub_char_scores
             double ub_word_scores
             double_array *best_char_scores
@@ -237,7 +239,7 @@ cdef class AdeftLongformScorer:
             for j in range(self.len_shortform):
                 if probe_results.char_prizes[j] > best_char_scores.array[j]:
                     ub_char_scores += probe_results.char_prizes[j]
-                else:
+                elif best_char_scores.array[j] > 0:
                     ub_char_scores += best_char_scores.array[j]
             if i > 1:
                 previous_word_scores = make_double_array(i)
@@ -258,11 +260,18 @@ cdef class AdeftLongformScorer:
             if upper_bound < best_score:
                 scores[i-1] = scores[i-2]*(W - w)/W
                 continue
-            perm_search(candidates_c, self.shortform_c,
-                        self.params_c,
-                        self.rho, i, results)
+            if upper_bound * self.rho < best_score:
+                opt_search(candidates_c, self.shortform_c,
+                           self.params_c, self.rho,
+                           i, 0, 0, results)
+            else:
+                max_inversions = get_max_inversions(best_score, upper_bound,
+                                                    self.rho)
+                opt_search(candidates_c, self.shortform_c,
+                           self.params_c, self.rho,
+                           i, 1, max_inversions, results)
             current_score = results.score
-            scores[i-1] = results.score
+            scores[i-1] = current_score
             if current_score >= best_score:
                 best_score = current_score
                 for j in range(self.len_shortform):
@@ -277,24 +286,35 @@ cdef class AdeftLongformScorer:
         return out
 
 
+cdef inline int get_max_inversions(double best_score,
+                                   double upper_bound,
+                                   double rho):
+    cdef int k = 0
+    if best_score <= 0.0:
+        return INT_MAX
+    while best_score <= upper_bound * cpow(rho, k):
+        k += 1
+    return k - 1
+
+
 cdef double opt_selection(double_array *word_prizes, int k):
     cdef:
-        int min_index, i, j
-        double min_value, temp, output
+        int max_index, i, j
+        double max_value, temp, output
     if k >= word_prizes.length:
         output = 0.0
         for i in range(word_prizes.length):
             output += word_prizes.array[i]
         return output
     for i in range(k):
-        min_index = i
-        min_value = word_prizes.array[i]
+        max_index = i
+        max_value = word_prizes.array[i]
         for j in range(i+i, word_prizes.length):
-            if word_prizes.array[j] < min_value:
-                min_index = j
-                min_value = word_prizes.array[j]
-                temp = word_prizes.array[min_index]
-                word_prizes.array[min_index] = word_prizes.array[i]
+            if word_prizes.array[j] > max_value:
[…]
