### 2019-10-03 598e05e
Reintroduce asserts to test
