### 2019-10-03 2cff45d
Update parameter descriptions in DocString
