### 2019-10-03 fbd5218
Describe default arg for word_scores
