### 2019-10-03 738ff6f
Add DocString for process_candidates
