### 2019-10-03 70cf67b
Add Documentation for get_score_resulst
