### 2019-10-03 eb87d18
Fix typo in opt_selection
