### 2019-10-03 a2bf6cc
Fix typo in process_candidates
