### 2019-10-03 8393ecc
Add additional DocStrings
