### 2019-10-03 65e264b
Add DocStrings for structs
