### 2019-10-03 c833e16
Make opt_search decide whether to permute
Decides not to permute if max_inversions = 0. Originally, whether or
not to permute was supplied as an argument
