### 2019-10-03 99d5edc
Remove extra newlines at end
