### 2019-10-03 ba9598f
Test only permutations with one inversion when ok
