### 2019-10-03 77af3a1
Update more DocStrings
