### 2019-10-03 aec3358
Get tests working again for make_candidates_array
