+++ adeft/tests/test_score.py
@@ -3,29 +3,33 @@ from adeft.tests.util import OptimizationTestCase, StitchTestCase, \
 
 
 def test_perm_search():
-    case1 = PermSearchTestCase(candidates=[[0], [0, 1], [0, 0, 1], [0, 0],
-                                           [1]],
-                               shortform=[1, 0],
-                               prizes=[[1.0], [1.0, 0.5], [1.0, 0.5, 0.25],
-                                       [1.0, 0.5], [1.0]],
+    case1 = PermSearchTestCase(candidates=[[1], [1, 0], [1, 1, 0], [1, 1],
+                                           [0]],
+                               shortform=[0, 1],
+                               indices=[[0], [0, 1], [0, 1, 2],
+                                       [0, 1], [0]],
                                penalties=[0.4, 0.2],
                                word_prizes=[1.0]*5,
                                word_penalties=[1.0, 2.0, 3.0, 4.0, 5.0],
-                               beta=0.5,
-                               rho=0.75,
-                               inv_penalty=0.9,
+                               alpha=0.5,
+                               beta=0.8,
+                               gamma=0.85,
+                               lambda_=0.75,
+                               rho=0.9,
                                len_perm=5,
                                result_score=(2/5)**(1/4) * (9/10))
     case2 = PermSearchTestCase(candidates=[[0, 1, 0], [1, 0, 0, 1]],
                                shortform=[0, 1],
-                               prizes=[[1.0, 1/8, 1/32],
-                                       [1.0, 1/2, 1/8, 1/64]],
+                               indices=[[0, 3, 5],
+                                       [0, 1, 3, 6]],
                                penalties=[1.0, 0.4],
                                word_prizes=[1.0, 1.0],
                                word_penalties=[1.0, 2.0],
-                               beta=0.55,
-                               rho=0.6,
-                               inv_penalty=0.9,
+                               alpha=0.5,
+                               beta=0.8,
+                               gamma=0.85,
+                               lambda_=0.6,
+                               rho=0.9,
                                len_perm=2,
                                result_score=1.0)
     for case in [case1, case2]:
@@ -34,29 +38,29 @@ def test_perm_search():
 
 def test_make_candidates_array():
     case1 = StitchTestCase(candidates=[[0], [0, 1], [1, 1, 0], [0, 0], [1]],
-                           prizes=[[1.], [1., 0.5], [1., 0.5, 0.25],
-                                   [1., 0.5], [1.]],
+                           indices=[[0], [0, 1], [0, 1, 2],
+                                   [0, 1], [0]],
                            word_prizes=[1., 0.9, 0.8, 0.7, 0.6],
                            W_array=[0.6, 1.3, 2.1, 3., 4.],
                            permutation=[2, 0, 1, 3, 4],
                            result_x=[-1, 1, -1, 1, -1, 0, -1, 0, -1, 0,
                                      -1, 1, -1, 0, -1, 0, -1, 1, -1],
-                           result_prizes=[0., 1., 0., 0.5, 0., 0.25,
-                                          0., 1., 0., 1., 0., 0.5, 0.,
-                                          1., 0., 0.5, 0., 1., 0.],
+                           result_indices=[-1, 0, -1, 1, -1, 2,
+                                          -1, 0, -1, 0, -1, 1, -1,
+                                           0, -1, 1, -1, 0, -1],
                            result_word_prizes=[0.8, 1., 0.9, 0.7, 0.6],
                            result_word_boundaries=[6, 8, 12, 16, 18])
 
     case2 = StitchTestCase(candidates=[[0], [0, 1], [1, 1, 0], [0, 0], [1]],
[…]
