### 2019-10-03 179ec05
Fix bug logical error in score calculations
