+++ adeft/score/score.pyx
@@ -569,8 +569,8 @@ cdef opt_input *make_opt_input(int n, int num_words):
     input_ = <opt_input *> PyMem_Malloc(sizeof(opt_input))
     input_.x = make_int_array(n)
     input_.indices = make_int_array(n)
-    input_.word_boundaries = <unsigned int *> \
-        PyMem_Malloc(num_words * sizeof(unsigned int))
+    input_.word_boundaries = <int *> \
+        PyMem_Malloc(num_words * sizeof(int))
     input_.word_prizes = make_double_array(num_words)
     return input_
 
@@ -818,11 +818,11 @@ cdef void optimize(opt_input *input_, opt_shortform *shortform,
         opt_results structure to where output is to be placed
     """
     cdef:
-        unsigned int n = input_.x.length
-        unsigned int m = shortform.y.length
+        int n = input_.x.length
+        int m = shortform.y.length
         double possibility1, possibility2, word_score, char_score, c, w
         double first_capture, prize
-        unsigned int i, j, k, first_capture_index
+        int i, j, k, current_fci
 
     # Dynamic initialization of score_lookup array and traceback pointer
     # array
@@ -836,6 +836,8 @@ cdef void optimize(opt_input *input_, opt_shortform *shortform,
         double **word_scores = (<double **>
                                 PyMem_Malloc((n+1) * sizeof(double *)))
         int **word_use = <int **> PyMem_Malloc((n+1) * sizeof(int *))
+        int **first_capture_index = <int **> \
+            PyMem_Malloc((n+1) * sizeof(int *))
         int **pointers = <int **> PyMem_Malloc(n * sizeof(int *))
     for i in range(n+1):
         score_lookup[i] = <double *> PyMem_Malloc((m+1) * sizeof(double))
@@ -843,6 +845,7 @@ cdef void optimize(opt_input *input_, opt_shortform *shortform,
         char_prizes[i] = <double *> PyMem_Malloc((m+1) * sizeof(double))
         word_scores[i] = <double *> PyMem_Malloc((m+1) * sizeof(double))
         word_use[i] = <int *> PyMem_Malloc((m+1) * sizeof(int))
+        first_capture_index[i] = <int *> PyMem_Malloc((m+1) * sizeof(int))
         if i != n:
             pointers[i] = <int *> PyMem_Malloc(m * sizeof(int))
     # Initialize lookup array
@@ -850,21 +853,23 @@ cdef void optimize(opt_input *input_, opt_shortform *shortform,
     char_scores[0][0] = 0
     word_scores[0][0] = 0
     word_use[0][0] = 0
+    first_capture_index[0][0] = -1
     for j in range(1, m+1):
         score_lookup[0][j] = -1e20
         char_scores[0][j] = 0
         word_scores[0][j] = 0
         word_use[0][j] = 0
+        first_capture_index[0][j] = -1
     for i in range(1, n+1):
         for j in range(0, m+1):
             score_lookup[i][j] = 0
             char_scores[i][j] = 0
             word_scores[i][j] = 0
             word_use[i][j] = 0
+            first_capture_index[i][j] = -1
     # Main loop
     k = 0
-    first_capture = 0
-    first_capture_index = -1
+    first_capture = 0.0
     for i in range(1, n+1):
         for j in range(1, m+1):
             # Case where element of x in current position matches
@@ -873,16 +878,19 @@ cdef void optimize(opt_input *input_, opt_shortform *shortform,
             if input_.x.array[i-1] == shortform.y.array[j-1]:
                 if word_use[i-1][j-1] == 0:
                     w = input_.word_prizes.array[k]
-                    first_capture_index = input_.indices.array[i-1]
-                    first_capture = cpow(params.alpha, first_capture_index)
[…]
