### 2019-10-03 eb90123
Get all score associated tests passing
