+++ adeft/tests/test_score.py
@@ -70,74 +70,76 @@ def test_make_candidates_array():
 
 
 def test_optimize():
-    # Two words with beginning characters matching the characters in the
-    # shortform
     case1 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1],
                                  y=[0, 1],
-                                 prizes=[0., 1., 0., 1., 0., 0.5, 0.],
+                                 indices=[-1, 0, -1, 0, -1, 1, -1],
                                  penalties=[0.4, 0.2],
                                  word_boundaries=[2, 6],
                                  word_prizes=[1, 1],
-                                 beta=0.5,
-                                 rho=0.75,
+                                 alpha=0.5,
+                                 beta=0.8,
+                                 gamma=0.85,
+                                 lambda_=0.75,
                                  W=2.,
                                  result_score=1.0,
                                  result_char_scores=[1., 1.])
 
-    # Single word containing consecutive characters from shortform
     case2 = OptimizationTestCase(x=[-1, 0, -1, 0, -1, 1, -1, 1, -1],
                                  y=[0, 1],
-                                 prizes=[0., 1., 0., 0.5, 0.,
-                                         0.25, 0., 0.125, 0.],
+                                 indices=[-1, 0, -1, 1, -1,
+                                          2, -1, 3, -1],
                                  penalties=[0.4, 0.2],
                                  word_boundaries=[8],
                                  word_prizes=[1],
-                                 beta=0.5,
-                                 rho=0.75,
+                                 alpha=0.5,
+                                 beta=0.8,
+                                 gamma=0.85,
+                                 lambda_=0.75,
                                  W=1.,
-                                 result_score=(3/4)**(3/4),
-                                 result_char_scores=[1.0, 0.5])
+                                 result_score=(1.8/2)**(3/4),
+                                 result_char_scores=[1.0, 0.8])
 
-    # Three words, shortform matches in two places. Highest scoring match
-    # has a larger word captured prize and smaller total letter prizes
     case3 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1, 0, -1, 1, -1],
                                  y=[0, 1],
-                                 prizes=[0., 1., 0., 1., 0., 1., 0., 0.5,
-                                         0., 0.25, 0.],
+                                 indices=[-1, 0, -1, 0, -1, 0, -1, 1,
+                                         -1, 2, -1],
                                  penalties=[0.4, 0.2],
                                  word_boundaries=[2, 4, 10],
                                  word_prizes=[0.5, 0.5, 2.0],
-                                 beta=0.5,
-                                 rho=0.5,
+                                 alpha=0.5,
+                                 beta=0.8,
+                                 gamma=0.85,
+                                 lambda_=0.5,
                                  W=3.0,
-                                 result_score=(3/4)**(1/2)*(2/3)**(1/2),
-                                 result_char_scores=[1.0, 0.5])
+                                 result_score=(1.8/2)**(1/2)*(2/3)**(1/2),
+                                 result_char_scores=[1.0, 0.8])
 
-    # Three words, shortform matches in two places. Highest scoring match
-    # has larger total letter prizes and smaller word captured prizes
     case4 = OptimizationTestCase(x=[-1, 0, -1, 1, -1, 0, -1, 0, -1, 1, -1],
                                  y=[0, 1],
-                                 prizes=[0., 1., 0., 1., 0., 1., 0., 0.5,
-                                         0., 0.25, 0.],
+                                 indices=[-1, 0, -1, 0, -1, 0, -1, 1,
+                                          -1, 2, -1],
[…]
