### 2019-10-04 31f831b
Make cython a dependency
