+++ setup.py
@@ -34,7 +34,7 @@ setup(name='adeft',
       ],
       packages=find_packages(),
       install_requires=['nltk', 'scikit-learn>=0.20.0', 'wget',
-                        'requests', 'flask'],
+                        'requests', 'flask', 'cython'],
       extras_require={'test': ['nose', 'coverage']},
       keywords=['nlp', 'biology', 'disambiguation'],
       ext_modules=cythonize(extensions,
