### 2019-10-04 5b8b641
Make sure Cython is installed before setup
