### 2019-10-08 23e320c
Add sentences about state of field
