+++ paper/paper.bib
@@ -53,4 +53,129 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    url={https://doi.org/10.15252/msb.20177651}
 }
 
+@article{wsd-survey2009,
+ author = {Navigli, Roberto},
+ title = {Word Sense Disambiguation: A Survey},
+ journal = {ACM Comput. Surv.},
+ issue_date = {February 2009},
+ volume = {41},
+ number = {2},
+ month = feb,
+ year = {2009},
+ issn = {0360-0300},
+ pages = {10:1--10:69},
+ articleno = {10},
+ numpages = {69},
+ url = {http://doi.acm.org/10.1145/1459352.1459355},
+ doi = {10.1145/1459352.1459355},
+ acmid = {1459355},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {WSD, Word sense disambiguation, lexical ambiguity, lexical semantics, semantic annotation, sense annotation, word sense discrimination},
+} 
+
+% 15952878 
+@Article{wsd-bio-overview2005,
+   Author="Schuemie, M. J.  and Kors, J. A.  and Mons, B. ",
+   Title="{{W}ord sense disambiguation in the biomedical domain: an overview}",
+   Journal="J. Comput. Biol.",
+   Year="2005",
+   Volume="12",
+   Number="5",
+   Pages="554--565",
+   Month="Jun"
+}
+
+% 24076369 
+@Article{wsd-difficulty2014,
+   Author="McInnes, B. T.  and Stevenson, M. ",
+   Title="{{D}etermining the difficulty of {W}ord {S}ense {D}isambiguation}",
+   Journal="J Biomed Inform",
+   Year="2014",
+   Volume="47",
+   Pages="83--90",
+   Month="Feb"
+}
+
+@inproceedings{disamb-bio2009,
+ author = {Stevenson, Mark and Guo, Yikun and Al Amri, Abdulaziz and Gaizauskas, Robert},
+ title = {Disambiguation of Biomedical Abbreviations},
+ booktitle = {Proceedings of the Workshop on Current Trends in Biomedical Natural Language Processing},
+ series = {BioNLP '09},
+ year = {2009},
+ isbn = {978-1-932432-30-5},
+ location = {Boulder, Colorado},
+ pages = {71--79},
+ numpages = {9},
+ url = {http://dl.acm.org/citation.cfm?id=1572364.1572374},
+ acmid = {1572374},
+ publisher = {Association for Computational Linguistics},
+ address = {Stroudsburg, PA, USA},
+} 
+
+[download] 
+
+% 20816855 
+@Article{disamb-bio2010,
+   Author="Stevenson, M.  and Guo, Y. ",
+   Title="{{D}isambiguation in the biomedical domain: the role of ambiguity type}",
+   Journal="J Biomed Inform",
+   Year="2010",
+   Volume="43",
+   Number="6",
+   Pages="972--981",
+   Month="Dec"
+}
+
+@inproceedings{deep-dive-wsd2018,
+    title = "A Deep Dive into Word Sense Disambiguation with {LSTM}",
+    author = "Le, Minh  and
+      Postma, Marten  and
+      Urbani, Jacopo  and
+      Vossen, Piek",
+    booktitle = "Proceedings of the 27th International Conference on Computational Linguistics",
+    month = aug,
+    year = "2018",
+    address = "Santa Fe, New Mexico, USA",
+    publisher = "Association for Computational Linguistics",
+    pages = "354--365",
+}
+
+@inproceedings{lmms-acl2019,
[…]
