### 2019-10-08 10c6eab
Change to correct format for citations
