+++ paper/paper.bib
@@ -113,7 +113,6 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
  address = {Stroudsburg, PA, USA},
 } 
 
-[download] 
 
 % 20816855 
 @Article{disamb-bio2010,
