### 2019-10-11 #1708 @johnbachman
@brainstorm @gbader We added a new paragraph to the manuscript briefly describing where Adeft fits in the context of word sense disambiguation more generally.
