### 2019-10-18 a848b69
Add list of stopwords to nlp
This addresses the issue that the default english stopwords in sklearn
have several issues as detailed in
J. Nothman, H. Qin and R. Yurchak (2018).
“Stop Word Lists in Free Open-source Software Packages”.
In Proc. Workshop for NLP Open Source Software.

nlp has been moved into its own subdirectory which includes the old
nlp.py as well as a json file containing stopwords
