### 2019-10-18 60db700
Get correct path to stopwords file
