### 2019-10-18 e3670e2
Add stopwords to MANIFEST.in and inlude in setup
