### 2019-10-18 4c8f48b
Make shortforms stopwords in AdeftClassifier
We have observed shortforms themselves becoming the most informative
feature for some disambiguation tasks. Making shortforms stopwords
solves this issue. This is a change that has been planned for some time.
