### 2019-10-18 966f2d9
Add shortform to stopwords in one more place
