### 2019-11-02 #1708 @gbader
The new paragraph looks great.  In my view, this paper should be accepted.

Sorry for the delay. I like the idea of this journal, but one issue with running peer review on github is that the emails all come from github, which I happen to have filtered due to the large number of automated emails I get from github.
