### 2019-11-08 4fcb3f8
Add new assertion to test
