### 2019-11-08 e06ed51
Keep module working when LongformScorer not built
