### 2019-11-08 7139b10
Change logging level to info
