+++ adeft/recognize.py
@@ -11,12 +11,12 @@ from adeft.nlp import tokenize, untokenize
 from adeft.util import get_candidate_fragments, get_candidate
 
 logger = logging.getLogger(__file__)
-    
+
 try:
     from adeft.score import AdeftLongformScorer
 except Exception:
-    logger.warning('OneShotRecognizer not available. AdeftLongformScorer'
-                   ' has not been built successfully.')
+    logger.info('OneShotRecognizer not available. AdeftLongformScorer'
+                ' has not been built successfully.')
 
 _stemmer = EnglishStemmer()
 
