### 2019-11-08 1d1dc05
Update README with instructions to build extensions
