+++ doc/conf.py
@@ -24,9 +24,9 @@ copyright = '2019, Albert Steppi'
 author = 'Albert Steppi'
 
 # The short X.Y version
-version = '0.4'
+version = '0.5'
 # The full version, including alpha/beta/rc tags
-release = '0.4.0'
+release = '0.5.0'
 
 
 # -- General configuration ---------------------------------------------------
