### 2019-11-08 537dbbf
Add cython to requirements in docs
