+++ doc/requirements.txt
@@ -5,3 +5,4 @@ nltk
 wget
 requests
 flask
+cython
