### 2019-11-08 08e823c
Merge all Cython extensions into a single extension
