+++ adeft/tests/util.pyx
@@ -1,183 +1,0 @@
-import logging
-
-from adeft.score.score cimport int_array, opt_results, candidates_array, \
-    opt_input, opt_params, opt_shortform
-from adeft.score.score cimport make_int_array, free_int_array, \
-    make_opt_results, free_opt_results, make_opt_input, free_opt_input, \
-    make_opt_params, free_opt_params, create_shortform, free_opt_shortform, \
-    make_candidates_array, free_candidates_array
-from adeft.score.score cimport optimize, stitch, opt_search
-
-logger = logging.getLogger(__name__)
-
-
-cdef class StitchTestCase:
-    """Test construction of candidates array and stitching"""
-    cdef:
-        list candidates, indices, word_prizes,
-        list permutation, result_x, result_word_prizes, result_indices
-        list result_word_boundaries, W_array
-    def __init__(self, candidates=None, indices=None,
-                 word_prizes=None,
-                 permutation=None, W_array=None,
-                 result_x=None, result_indices=None, result_word_prizes=None,
-                 result_word_boundaries=None):
-        self.candidates = candidates
-        self.indices = indices
-        self.word_prizes = word_prizes
-        self.W_array = W_array
-        self.permutation = permutation
-        self.result_x = result_x
-        self.result_indices = result_indices
-        self.result_word_prizes = result_word_prizes
-        self.result_word_boundaries = result_word_boundaries
-
-    def run_test(self):
-        cdef:
-            opt_input *input_
-            int_array *perm
-            candidates_array *candidates
-
-        candidates = make_candidates_array(self.candidates,
-                                           self.indices,
-                                           self.word_prizes,
-                                           self.W_array)
-        n = len(self.permutation)
-        perm = make_int_array(n)
-        for i in range(n):
-            perm.array[i] = self.permutation[i]
-        total_length = candidates.cum_lengths[n - 1]
-        input_ = make_opt_input(2*total_length + 1, n)
-        stitch(candidates, perm.array, n, input_)
-        x, ind, wp, wb = [], [], [], []
-        length = input_.x.length
-        for i in range(length):
-            x.append(input_.x.array[i])
-            ind.append(input_.indices.array[i])
-        for j in range(input_.word_prizes.length):
-            wp.append(input_.word_prizes.array[j])
-            wb.append(input_.word_boundaries[j])
-        free_candidates_array(candidates)
-        free_opt_input(input_)
-        free_int_array(perm)
-        assert x == self.result_x
-        assert ind == self.result_indices
-        assert wp == self.result_word_prizes
-        assert wb == self.result_word_boundaries
-
-
-cdef class PermSearchTestCase:
-    cdef:
-        list shortform, candidates, indices, penalties, word_prizes
-        list word_penalties
-        double alpha, beta, gamma, lambda_, rho, result_score
-        int len_perm
-    def __init__(self, shortform=None, candidates=None, indices=None,
-                 penalties=None, word_prizes=None, word_penalties=None,
-                 alpha=None, beta=None, gamma=None, lambda_=None, rho=None,
-                 len_perm=None,
-                 result_score=None):
-        self.shortform = shortform
-        self.candidates = candidates
-        self.indices = indices
-        self.penalties = penalties
-        self.word_prizes = word_prizes
-        self.word_penalties = word_penalties
-        self.alpha = alpha
-        self.beta = beta
-        self.gamma = gamma
-        self.lambda_ = lambda_
-        self.rho = rho
-        self.len_perm = len_perm
-        self.result_score = result_score
-
-    def run_test(self):
-        cdef:
-            candidates_array *candidates
-            opt_shortform *shortform
-            opt_params *params
-            opt_results *results
[…]
