+++ adeft/score/score.pyx
@@ -1,3 +1,6 @@
+import logging
+
+
 from libc.math cimport pow as cpow
 cdef extern from 'limits.h':
     int INT_MAX
@@ -5,8 +8,55 @@ from cython cimport boundscheck, wraparound
 from cpython.mem cimport PyMem_Malloc, PyMem_Free
 
 from adeft.nlp import stopwords
-from adeft.score.permutations cimport permuter, make_permuter, \
-    free_permuter, update_permuter
+
+include 'permutations.pyx'
+
+logger = logging.getLogger(__name__)
+
+
+cdef struct int_array:
+    int *array
+    int length
+
+
+cdef struct double_array:
+    double *array
+    int length
+
+    
+cdef struct opt_results:
+    double score
+    double *char_prizes
+
+    
+cdef struct candidates_array:
+    int_array **array
+    int_array **indices
+    double *word_prizes
+    double *W_array
+    int *cum_lengths
+    int length
+
+
+cdef struct opt_input:
+    int_array *x
+    int_array *indices
+    int *word_boundaries
+    double_array *word_prizes
+    double W
+
+
+cdef struct opt_params:
+    double alpha, beta, gamma, lambda_
+
+
+cdef struct opt_shortform:
+    int_array *y
+    double_array *penalties
+
+
+cdef struct perm_out:
+    double score
 
 
 cdef class AdeftLongformScorer:
@@ -1005,3 +1055,175 @@ cdef void optimize(opt_input *input_, opt_shortform *shortform,
     PyMem_Free(char_prizes)
     PyMem_Free(word_use)
     PyMem_Free(pointers)
+
+
+cdef class StitchTestCase:
+    """Test construction of candidates array and stitching"""
+    cdef:
+        list candidates, indices, word_prizes,
+        list permutation, result_x, result_word_prizes, result_indices
+        list result_word_boundaries, W_array
+    def __init__(self, candidates=None, indices=None,
+                 word_prizes=None,
+                 permutation=None, W_array=None,
+                 result_x=None, result_indices=None, result_word_prizes=None,
+                 result_word_boundaries=None):
+        self.candidates = candidates
+        self.indices = indices
+        self.word_prizes = word_prizes
+        self.W_array = W_array
+        self.permutation = permutation
+        self.result_x = result_x
+        self.result_indices = result_indices
+        self.result_word_prizes = result_word_prizes
+        self.result_word_boundaries = result_word_boundaries
+
+    def run_test(self):
+        cdef:
+            opt_input *input_
+            int_array *perm
+            candidates_array *candidates
+
+        candidates = make_candidates_array(self.candidates,
+                                           self.indices,
+                                           self.word_prizes,
+                                           self.W_array)
+        n = len(self.permutation)
+        perm = make_int_array(n)
+        for i in range(n):
+            perm.array[i] = self.permutation[i]
+        total_length = candidates.cum_lengths[n - 1]
+        input_ = make_opt_input(2*total_length + 1, n)
+        stitch(candidates, perm.array, n, input_)
+        x, ind, wp, wb = [], [], [], []
+        length = input_.x.length
+        for i in range(length):
+            x.append(input_.x.array[i])
+            ind.append(input_.indices.array[i])
+        for j in range(input_.word_prizes.length):
+            wp.append(input_.word_prizes.array[j])
+            wb.append(input_.word_boundaries[j])
+        free_candidates_array(candidates)
+        free_opt_input(input_)
+        free_int_array(perm)
+        assert x == self.result_x
+        assert ind == self.result_indices
+        assert wp == self.result_word_prizes
+        assert wb == self.result_word_boundaries
+
+
+cdef class PermSearchTestCase:
+    cdef:
[…]
