### 2019-11-08 c700723
Remove Cython from Doc requirements
