+++ setup.py
@@ -28,13 +28,13 @@ if USE_CYTHON:
                            compiler_directives={'language_level': 3})
 
 setup(name='adeft',
-      version='0.5.0',
+      version='0.5.1',
       description=('Acromine based Disambiguation of Entities From'
                    ' Text'),
       long_description=long_description,
       long_description_content_type='text/markdown',
       url='https://github.com/indralab/adeft',
-      download_url='https://github.com/indralab/adeft/archive/0.5.0.tar.gz',
+      download_url='https://github.com/indralab/adeft/archive/0.5.1.tar.gz',
       author='adeft developers, Harvard Medical School',
       author_email='albert_steppi@hms.harvard.edu',
       classifiers=[
