### 2019-12-16 dcfb46f
Increase possible length of grounding strings
