+++ adeft/modeling/classify.py
@@ -294,7 +294,7 @@ def load_model_info(model_info):
     tfidf.vocabulary_ = model_info['tfidf']['vocabulary_']
     tfidf.idf_ = model_info['tfidf']['idf_']
     logit.classes_ = np.array(model_info['logit']['classes_'],
-                              dtype='<U32')
+                              dtype='<U64')
     logit.intercept_ = np.array(model_info['logit']['intercept_'])
     logit.coef_ = np.array(model_info['logit']['coef_'])
 
