### 2019-12-19 f707726
Add more information to DocString
