### 2019-12-19 6612f5c
Add func to load disambiguator directly from its path
