### 2019-12-19 2cb931d
Add DocString for func to load disambiguator directly
