### 2020-01-10 #1708 @GullyAPCBurns
I was able to review the paper and software. The code looks great and the documentation is exemplary. I ran the available Jupyter notebooks and was able to disambiguate 'Gut Cancer (GC)' effectively.

Again, I apologize for not having done this immediately. I'd recommend that you proceed with publication 'as is'.
