### 2020-01-10 #1708 @brainstorm
Looks good to my too, @johnbachman, next step is for you to deposit it on Zenodo and we'll archive it in JOSS, let me know if you run into any problems with that. Pretty close to publication now ;)
