### 2020-01-13 7a1dc5f
Replace gridsearch object with cv_results dict
