### 2020-01-13 5d8496f
Implement method for feature importance scores
