### 2020-01-14 8f4abfc
Bump version to 0.5.2
