### 2020-01-14 #1708 @johnbachman
Hi @brainstorm, the release for the paper is up at Zenodo here: https://zenodo.org/record/3608205

The version is 0.5.2, DOI is 10.5281/zenodo.3608205.

Also, a question--should we add in DOIs for the 7 listed as "missing"?
