### 2020-01-14 654e4a4
Bump version in all locations to 0.5.3
