+++ adeft/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.5.1'
+__version__ = '0.5.3'
 
 from adeft.download import get_available_models
 
