### 2020-01-14 #1708 @brainstorm
No worries John, gotcha. @danielskatz LGTM, can you find any irregularity pre-publishing it for good?
