### 2020-01-14 7c8ad00
Add test of feature importance for binary case
