### 2020-01-14 fbadca2
Get feature importance working for binary case
