### 2020-01-14 25c490a
Add new test conditions
