### 2020-01-14 e03ff20
Fix mistake in which class label to use
An explanatory comment has been added
