### 2020-01-14 2aeee4e
Add checks that importance scores make sense
