### 2020-01-14 574d18c
Revert to storing grid_search instead of cv_results
An overhaul of the reporting of model statistics is coming in the
future. This will be left alone until that time comes
