### 2020-01-14 0b06e2d
Make feature importace info serializable
