### 2020-01-14 73e35c7
Add test for serialization of feature importance scores
The tests should probably be made more granular. Right now many
separate tests are grouped together in the same function
