### 2020-01-15 #1708 @danielskatz
👋 @openjournals/joss-eics - I think this may be ready for you
