### 2020-01-15 0d8bd0b
Update version numbers in setup, __init__, doc
