+++ doc/conf.py
@@ -26,7 +26,7 @@ author = 'Albert Steppi'
 # The short X.Y version
 version = '0.5'
 # The full version, including alpha/beta/rc tags
-release = '0.5.3'
+release = '0.5.4'
 
 
 # -- General configuration ---------------------------------------------------
