### 2020-01-15 f5799e3
Update test for change in stopword handling
