### 2020-01-15 42fdc20
Remove extraneous whitespace
