### 2020-01-15 #1708 @kyleniemeyer
Hi @johnbachman, can you update the Zenodo archive title to match the paper? Also, it looks like some of your references may be missing DOIs 👆, if you could add those (you may want to check that the suggestions are correct first).
