### 2020-01-15 c35049d
Bump version number to 0.5.5
