+++ adeft/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.5.4'
+__version__ = '0.5.5'
 
 from adeft.download import get_available_models
 
