### 2020-01-16 1e0b57e
Add DOIs for references
