+++ paper/paper.bib
@@ -72,7 +72,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
  publisher = {ACM},
  address = {New York, NY, USA},
  keywords = {WSD, Word sense disambiguation, lexical ambiguity, lexical semantics, semantic annotation, sense annotation, word sense discrimination},
-} 
+}
 
 % 15952878 
 @Article{wsd-bio-overview2005,
@@ -83,7 +83,8 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Volume="12",
    Number="5",
    Pages="554--565",
-   Month="Jun"
+   Month="Jun",
+   doi={10.1089/cmb.2005.12.554},
 }
 
 % 24076369 
@@ -94,7 +95,8 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Year="2014",
    Volume="47",
    Pages="83--90",
-   Month="Feb"
+   Month="Feb",
+   doi={10.1016/j.jbi.2013.09.009},
 }
 
 @inproceedings{disamb-bio2009,
@@ -111,6 +113,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
  acmid = {1572374},
  publisher = {Association for Computational Linguistics},
  address = {Stroudsburg, PA, USA},
+ doi={10.3115/1572364.1572374},
 } 
 
 
@@ -123,7 +126,8 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Volume="43",
    Number="6",
    Pages="972--981",
-   Month="Dec"
+   Month="Dec",
+   doi={10.1016/j.jbi.2010.08.009},
 }
 
 @inproceedings{deep-dive-wsd2018,
@@ -145,7 +149,8 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
   author={Daniel Loureiro and Al{\'\i}pio M{\'a}rio Jorge},
   booktitle={Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics},
   pages={forthcoming},
-  year={2019}
+  year={2019},
+  doi={10.18653/v1/p19-1569},
 }
 
 @InProceedings{glosses-nwsd,
@@ -159,7 +164,8 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
   year = 	"2018",
   publisher = 	"Association for Computational Linguistics",
   pages = 	"2473--2482",
-  url = 	"http://aclweb.org/anthology/P18-1230"
+  url = 	"http://aclweb.org/anthology/P18-1230",
+  doi={https://doi.org/10.18653/v1/p18-1230},
 }
 
 % 15064284 
@@ -170,7 +176,8 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Year="2004",
    Volume="11",
    Number="4",
-   Pages="320--331"
+   Pages="320--331",
+   doi={10.1197/jamia.m1533},
 }
 
 
