### 2020-01-16 0441dce
Add URLs in addition to DOIs
