+++ paper/paper.bib
@@ -85,6 +85,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Pages="554--565",
    Month="Jun",
    doi={10.1089/cmb.2005.12.554},
+   url={https://doi.org/10.1089/cmb.2005.12.554},
 }
 
 % 24076369 
@@ -97,6 +98,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Pages="83--90",
    Month="Feb",
    doi={10.1016/j.jbi.2013.09.009},
+   url={https://doi.org/10.1016/j.jbi.2013.09.009},
 }
 
 @inproceedings{disamb-bio2009,
@@ -114,6 +116,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
  publisher = {Association for Computational Linguistics},
  address = {Stroudsburg, PA, USA},
  doi={10.3115/1572364.1572374},
+ url={https://doi.org/10.3115/1572364.1572374},
 } 
 
 
@@ -128,6 +131,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Pages="972--981",
    Month="Dec",
    doi={10.1016/j.jbi.2010.08.009},
+   url={https://doi.org/10.1016/j.jbi.2010.08.009},
 }
 
 @inproceedings{deep-dive-wsd2018,
@@ -151,6 +155,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
   pages={forthcoming},
   year={2019},
   doi={10.18653/v1/p19-1569},
+  url={https://doi.org/10.18653/v1/p19-1569},
 }
 
 @InProceedings{glosses-nwsd,
@@ -178,6 +183,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
    Number="4",
    Pages="320--331",
    doi={10.1197/jamia.m1533},
+   url={https://doi.org/10.1197/jamia.m1533},
 }
 
 
