### 2020-01-16 #1708 @kyleniemeyer
@johnbachman your earlier requests to check the references didn't work because the PDF needed to be rebuilt first—thanks for adding those!

Can you remove `https://doi.org/` from that one DOI causing an error? That should be the last edit needed.
