+++ paper/paper.bib
@@ -170,7 +170,7 @@ url={http://dl.acm.org/citation.cfm?id=1953048.2078195}
   publisher = 	"Association for Computational Linguistics",
   pages = 	"2473--2482",
   url = 	"http://aclweb.org/anthology/P18-1230",
-  doi={https://doi.org/10.18653/v1/p18-1230},
+  doi={10.18653/v1/p18-1230},
 }
 
 % 15064284 
