### 2020-01-16 #1708 @johnbachman
Will do. After that's done, I will tag and release the latest version and update the Zenodo link.
