### 2020-01-16 #1708 @johnbachman
@brainstorm @kyleniemeyer The updated Zenodo archive (for Adeft version 5.5) is 10.5281/zenodo.3610528. Once that is updated we should be good to go.
